Split determinism tests into separate modules under tests/deterministic
