--- warp/tests/test_deterministic.py
@@ -1,3822 +1,0 @@
-# SPDX-FileCopyrightText: Copyright (c) 2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
-# SPDX-License-Identifier: Apache-2.0
-
-"""Tests for deterministic execution mode.
-
-Validates that deterministic modes produce bit-exact reproducible results for
-atomic operations across multiple runs.
-"""
-
-import gc
-import re
-import sys
-import unittest
-from pathlib import Path
-
-import numpy as np
-
-import warp as wp
-from warp._src import deterministic as wp_deterministic
-from warp._src.types import _np_bfloat16_bits_to_float32
-from warp.tests.unittest_utils import *
-
-
-def _reference_scatter_add_float32(data_np, indices_np, out_size):
-    """Compute the canonical left-to-right float32 scatter reduction on CPU."""
-    expected = np.zeros(out_size, dtype=np.float32)
-    for value, index in zip(data_np, indices_np, strict=True):
-        expected[index] = np.float32(expected[index] + value)
-    return expected
-
-
-def _bfloat16_numpy_bits(arr):
-    """Return raw bfloat16 bits from a Warp bfloat16 ``.numpy()`` result."""
-    arr = np.asarray(arr)
-    if arr.dtype == np.uint16:
-        return arr.copy()
-    return arr.view(np.uint16).copy()
-
-
-def _bfloat16_numpy_to_float32(arr):
-    """Convert a Warp bfloat16 ``.numpy()`` result to float32."""
-    arr = np.asarray(arr)
-    if arr.dtype == np.uint16:
-        return _np_bfloat16_bits_to_float32(arr)
-    return arr.astype(np.float32)
-
-
-def _set_test_module_options(options):
-    wp.set_module_options(options, module=sys.modules[__name__])
-
-
-def _get_test_module_options():
-    return wp.get_module_options(module=sys.modules[__name__])
-
-
-# ---------------------------------------------------------------------------
-# Scatter/reduce kernels: return value unused.
-# ---------------------------------------------------------------------------
-
-
-@wp.kernel
-def scatter_add_kernel(
-    data: wp.array[wp.float32],
-    dest_indices: wp.array[wp.int32],
-    output: wp.array[wp.float32],
-):
-    """Each thread atomically adds to output[dest_indices[tid]]."""
-    tid = wp.tid()
-    idx = dest_indices[tid]
-    wp.atomic_add(output, idx, data[tid])
-
-
-@wp.kernel
-def augassign_add_kernel(
-    data: wp.array[wp.float32],
-    dest_indices: wp.array[wp.int32],
-    output: wp.array[wp.float32],
-):
-    """Same as scatter_add_kernel but using += syntax."""
-    tid = wp.tid()
-    idx = dest_indices[tid]
-    output[idx] += data[tid]
-
-
-@wp.kernel
-def multi_array_atomic_kernel(
-    data: wp.array[wp.float32],
-    dest_indices: wp.array[wp.int32],
-    out_a: wp.array[wp.float32],
-    out_b: wp.array[wp.float32],
-    out_c: wp.array[wp.float32],
-):
-    """Atomic add to three different output arrays from the same kernel."""
-    tid = wp.tid()
-    idx = dest_indices[tid]
-    val = data[tid]
-    wp.atomic_add(out_a, idx, val)
-    wp.atomic_add(out_b, idx, val * 2.0)
-    wp.atomic_add(out_c, idx, val * 3.0)
-
-
-@wp.kernel
-def atomic_sub_kernel(
-    data: wp.array[wp.float32],
-    dest_indices: wp.array[wp.int32],
-    output: wp.array[wp.float32],
-):
-    """Atomic sub test."""
-    tid = wp.tid()
-    idx = dest_indices[tid]
-    wp.atomic_sub(output, idx, data[tid])
-
-
-@wp.kernel
-def atomic_add_2d_kernel(
-    data: wp.array[wp.float32],
-    row_indices: wp.array[wp.int32],
-    col_indices: wp.array[wp.int32],
-    output: wp.array2d[wp.float32],
-):
-    """Atomic add to a 2D array."""
-    tid = wp.tid()
-    r = row_indices[tid]
-    c = col_indices[tid]
-    wp.atomic_add(output, r, c, data[tid])
-
-
-@wp.kernel
-def sliced_2d_atomic_add_kernel(
-    data: wp.array[wp.float32],
-    row_indices: wp.array[wp.int32],
-    col_indices: wp.array[wp.int32],
-    output: wp.array2d[wp.float32],
-):
-    """Atomic add through a sliced ``output[row]`` view."""
-    tid = wp.tid()
-    row = row_indices[tid]
-    col = col_indices[tid]
-    wp.atomic_add(output[row], col, data[tid])
-
-
-@wp.kernel
-def sliced_3d_atomic_add_kernel(
-    data: wp.array[wp.float32],
-    row_indices: wp.array[wp.int32],
-    col_indices: wp.array[wp.int32],
-    depth_indices: wp.array[wp.int32],
-    output: wp.array3d[wp.float32],
-):
-    """Atomic add through a sliced ``output[row, col]`` view."""
-    tid = wp.tid()
-    row = row_indices[tid]
-    col = col_indices[tid]
-    depth = depth_indices[tid]
-    wp.atomic_add(output[row, col], depth, data[tid])
-
-
-@wp.kernel
-def atomic_half_kernel(
-    data: wp.array[wp.float16],
-    dest_indices: wp.array[wp.int32],
-    output: wp.array[wp.float16],
-):
-    """Atomic add with float16."""
-    tid = wp.tid()
-    idx = dest_indices[tid]
-    wp.atomic_add(output, idx, data[tid])
-
-
-@wp.kernel
-def atomic_bfloat16_kernel(
-    data: wp.array[wp.bfloat16],
-    dest_indices: wp.array[wp.int32],
-    output: wp.array[wp.bfloat16],
-):
-    """Atomic add with bfloat16."""
-    tid = wp.tid()
-    idx = dest_indices[tid]
-    wp.atomic_add(output, idx, data[tid])
-
-
-@wp.kernel
-def atomic_bfloat16_minmax_kernel(
-    data: wp.array[wp.bfloat16],
-    dest_indices: wp.array[wp.int32],
-    output_min: wp.array[wp.bfloat16],
-    output_max: wp.array[wp.bfloat16],
-):
-    """Atomic min/max with bfloat16."""
-    tid = wp.tid()
-    idx = dest_indices[tid]
-    wp.atomic_min(output_min, idx, data[tid])
-    wp.atomic_max(output_max, idx, data[tid])
-
-
-@wp.kernel
-def atomic_double_kernel(
-    data: wp.array[wp.float64],
-    dest_indices: wp.array[wp.int32],
-    output: wp.array[wp.float64],
-):
-    """Atomic add with float64."""
-    tid = wp.tid()
-    idx = dest_indices[tid]
-    wp.atomic_add(output, idx, data[tid])
-
-
-@wp.kernel
-def vec3_scatter_add_kernel(
-    data: wp.array[wp.vec3],
-    dest_indices: wp.array[wp.int32],
-    output: wp.array[wp.vec3],
-):
-    """Atomic add with ``wp.vec3`` values."""
-    tid = wp.tid()
-    wp.atomic_add(output, dest_indices[tid], data[tid])
-
-
-@wp.kernel
-def gather_address_kernel(
-    values: wp.array[wp.float32],
-    source_indices: wp.array[wp.int32],
-    output: wp.array[wp.float32],
-):
-    """Read many output lanes from the same input lanes.
-
-    The forward pass is a plain gather, but the generated backward pass scatters
-    output gradients back into ``values.grad`` with atomic additions.
-    """
-    tid = wp.tid()
-    output[tid] = values[source_indices[tid]]
-
-
-@wp.kernel
-def vec3_gather_address_kernel(
-    values: wp.array[wp.vec3],
-    source_indices: wp.array[wp.int32],
-    output: wp.array[wp.vec3],
-):
-    """Vector-valued gather covering cloth/particle-style position gradients."""
-    tid = wp.tid()
-    output[tid] = values[source_indices[tid]]
-
-
-@wp.kernel
-def gather_with_nongrad_input_kernel(
-    values: wp.array[wp.float32],
-    source_indices: wp.array[wp.int32],
-    scale: wp.array[wp.float32],
-    output: wp.array[wp.float32],
-):
-    """Read a non-differentiable input while accumulating into another gradient."""
-    tid = wp.tid()
-    output[tid] = values[source_indices[tid]] * scale[0]
-
-
-@wp.kernel
-def vec3_atomic_minmax_kernel(
-    points: wp.array[wp.vec3],
-    out_min: wp.array[wp.vec3],
-    out_max: wp.array[wp.vec3],
-):
-    """Component-wise deterministic min/max for bounding-box style reductions."""
-    tid = wp.tid()
-    p = points[tid]
-    wp.atomic_min(out_min, 0, p)
-    wp.atomic_max(out_max, 0, p)
-
-
-@wp.kernel
-def mat33_scatter_add_kernel(
-    data: wp.array[wp.mat33],
-    dest_indices: wp.array[wp.int32],
-    output: wp.array[wp.mat33],
-):
-    """Atomic add with ``wp.mat33`` values."""
-    tid = wp.tid()
-    wp.atomic_add(output, dest_indices[tid], data[tid])
-
-
-@wp.func
-def _det_closure_transform_a(x: wp.float32) -> wp.float32:
-    return x + wp.float32(1.0)
-
-
-@wp.func
-def _det_closure_transform_b(x: wp.float32) -> wp.float32:
-    return x + wp.float32(2.0)
-
-
-@wp.func
-def _det_func_scatter_add_leaf(arr: wp.array[wp.float32], idx: int, value: wp.float32):
-    wp.atomic_add(arr, idx, value)
-
-
-@wp.func
-def _det_func_scatter_add_wrapper(dst: wp.array[wp.float32], idx: int, value: wp.float32):
-    _det_func_scatter_add_leaf(dst, idx, value)
-
-
-@wp.struct
-class _DetStructCounterWriter:
-    counter: wp.array[wp.int32]
-    output: wp.array[wp.float32]
-
-
-@wp.struct
-class _DetNameCollisionStruct:
-    b: wp.array[wp.float32]
-
-
-@wp.struct
-class _DetTileStackEntry:
-    value: wp.int32
-    weight: wp.float32
-
-
-@wp.func
-def _det_struct_counter_write(writer: _DetStructCounterWriter, value: wp.float32):
-    slot = wp.atomic_add(writer.counter, 0, 1)
-    writer.output[slot] = value
-
-
-@wp.func
-def _det_counter_write(counter: wp.array[wp.int32], output: wp.array[wp.float32], value: wp.float32):
-    slot = wp.atomic_add(counter, 0, 1)
-    output[slot] = value
-
-
-@wp.func
-def _det_set_int_scratch(scratch: wp.array[wp.int32], index: int, value: int):
-    scratch[index] = value
-
-
-@wp.func
-def _det_increment_array(output: wp.array[wp.float32], index: int):
-    output[index] = output[index] + 1.0
-
-
-def _make_deterministic_closure_kernel(transform_func):
-    @wp.kernel(module="unique", module_options={"deterministic": "run_to_run"})
-    def _deterministic_closure_kernel(
-        data: wp.array[wp.float32],
-        output: wp.array[wp.float32],
-    ):
-        tid = wp.tid()
-        wp.atomic_add(output, tid % 8, transform_func(data[tid]))
-
-    return _deterministic_closure_kernel
-
-
-@wp.kernel(module="unique", module_options={"deterministic": "run_to_run"})
-def struct_field_counter_kernel(
-    data: wp.array[wp.float32],
-    counts: wp.array[wp.int32],
-    writer: _DetStructCounterWriter,
-):
-    tid = wp.tid()
-    count = counts[tid]
-    if count > 0:
-        base = wp.atomic_add(writer.counter, 0, count)
-        for i in range(count):
-            writer.output[base + i] = data[tid] + wp.float32(i) * wp.float32(0.5)
-
-
-@wp.kernel(module="unique", module_options={"deterministic": "run_to_run"})
-def struct_field_helper_counter_kernel(
-    data: wp.array[wp.float32],
-    flags: wp.array[wp.int32],
-    writer: _DetStructCounterWriter,
-):
-    tid = wp.tid()
-    if flags[tid] != 0:
-        _det_struct_counter_write(writer, data[tid])
-
-
-@wp.kernel(module="unique", module_options={"deterministic": "run_to_run"})
-def struct_tile_stack_kernel(out: wp.array[wp.int32]):
-    _i, j = wp.tid()
-    stack = wp.tile_stack(capacity=16, dtype=_DetTileStackEntry)
-    entry = _DetTileStackEntry()
-    entry.value = j
-    entry.weight = wp.float32(j) * wp.float32(0.25)
-    wp.tile_stack_push(stack, entry, j < 8)
-    popped, slot = wp.tile_stack_pop(stack)
-    if slot >= 0:
-        out[slot] = popped.value
-
-
-@wp.kernel(module="unique", module_options={"deterministic": "run_to_run"})
-def helper_counter_side_effect_kernel(
-    counter: wp.array[wp.int32],
-    output: wp.array[wp.float32],
-    scratch: wp.array[wp.float32],
-):
-    """Normal stores before helper counter calls must be suppressed in phase 0."""
-    tid = wp.tid()
-    scratch[tid] = scratch[tid] + 1.0
-    _det_counter_write(counter, output, float(tid))
-
-
-@wp.kernel(module="unique", module_options={"deterministic": "run_to_run"})
-def counter_with_helper_store_kernel(
-    counter: wp.array[wp.int32],
-    output: wp.array[wp.float32],
-    scratch: wp.array[wp.float32],
-):
-    """Pure write helpers called from counter kernels must be skipped in phase 0."""
-    tid = wp.tid()
-    slot = wp.atomic_add(counter, 0, 1)
-    _det_increment_array(scratch, tid)
-    output[slot] = float(tid)
-
-
-@wp.kernel(module="unique", module_options={"deterministic": "run_to_run"})
-def helper_name_collision_kernel(
-    a_b: wp.array[wp.float32],
-    a: _DetNameCollisionStruct,
-    data: wp.array[wp.float32],
-):
-    tid = wp.tid()
-    wp.atomic_add(a_b, 0, data[tid])
-    wp.atomic_add(a.b, 0, data[tid] * wp.float32(2.0))
-
-
-@wp.kernel
-def func_scatter_add_kernel(
-    data: wp.array[wp.float32],
-    dest_indices: wp.array[wp.int32],
-    output: wp.array[wp.float32],
-):
-    tid = wp.tid()
-    _det_func_scatter_add_leaf(output, dest_indices[tid], data[tid])
-
-
-@wp.kernel
-def nested_func_scatter_add_kernel(
-    data: wp.array[wp.float32],
-    dest_indices: wp.array[wp.int32],
-    accum: wp.array[wp.float32],
-):
-    tid = wp.tid()
-    _det_func_scatter_add_wrapper(accum, dest_indices[tid], data[tid])
-
-
-@wp.kernel
-def triple_scatter_add_kernel(
-    data: wp.array[wp.float32],
-    output: wp.array[wp.float32],
-):
-    """Emit three deterministic scatter records per thread to the same target."""
-    tid = wp.tid()
-    val = data[tid]
-    wp.atomic_add(output, 0, val)
-    wp.atomic_add(output, 0, val * 2.0)
-    wp.atomic_add(output, 0, val * 3.0)
-
-
-@wp.kernel(module="unique", module_options={"deterministic": "run_to_run", "deterministic_max_records": 4})
-def loop_scatter_add_kernel(
-    data: wp.array[wp.float32],
-    counts: wp.array[wp.int32],
-    output: wp.array[wp.float32],
-):
-    """Emit a data-dependent number of scatter records to the same target."""
-    tid = wp.tid()
-    val = data[tid]
-    count = counts[tid]
-    for _ in range(count):
-        wp.atomic_add(output, 0, val)
-
-
-@wp.kernel(module="unique", module_options={"deterministic": "run_to_run", "deterministic_max_records": 1})
-def underprovisioned_loop_scatter_kernel(
-    data: wp.array[wp.float32],
-    counts: wp.array[wp.int32],
-    output: wp.array[wp.float32],
-):
-    tid = wp.tid()
-    count = counts[tid]
-    for _ in range(count):
-        wp.atomic_add(output, 0, data[tid])
-
-
-@wp.kernel(module="unique")
-def mixed_reduce_op_same_array_kernel(
-    data: wp.array[wp.float32],
-    output: wp.array[wp.float32],
-):
-    """Apply different atomic reductions to the same destination array."""
-    tid = wp.tid()
-    wp.atomic_add(output, 0, data[tid])
-    wp.atomic_max(output, 0, 1.0)
-
-
-# ---------------------------------------------------------------------------
-# Counter kernels: return value used.
-# ---------------------------------------------------------------------------
-
-
-@wp.kernel
-def counter_kernel(
-    data: wp.array[wp.float32],
-    counter: wp.array[wp.int32],
-    output: wp.array[wp.float32],
-):
-    """Allocate a slot and write data to it."""
-    tid = wp.tid()
-    slot = wp.atomic_add(counter, 0, 1)
-    output[slot] = data[tid]
-
-
-@wp.kernel
-def conditional_counter_kernel(
-    data: wp.array[wp.float32],
-    threshold: wp.float32,
-    counter: wp.array[wp.int32],
-    output: wp.array[wp.float32],
-):
-    """Stream compaction: only emit elements above threshold."""
-    tid = wp.tid()
-    val = data[tid]
-    if val > threshold:
-        slot = wp.atomic_add(counter, 0, 1)
-        output[slot] = val
-
-
-@wp.kernel
-def variable_counter_kernel(
-    counts: wp.array[wp.int32],
-    counter: wp.array[wp.int32],
-    output: wp.array[wp.int32],
-):
-    """Reserve a variable number of slots per thread."""
-    tid = wp.tid()
-    count = counts[tid]
-    slot = wp.atomic_add(counter, 0, count)
-
-    for i in range(count):
-        output[slot + i] = tid * 10 + i
-
-
-@wp.kernel
-def static_index_counter_kernel(
-    counter: wp.array[wp.int32],
-    output: wp.array[wp.int32],
-):
-    """Reserve slots from a fixed nonzero counter index."""
-    tid = wp.tid()
-    slot = wp.atomic_add(counter, 1, 1)
-    output[slot] = tid
-
-
-@wp.kernel
-def indexed_counter_kernel(
-    values: wp.array[wp.int32],
-    bins: wp.array[wp.int32],
-    counters: wp.array[wp.int32],
-    output: wp.array2d[wp.int32],
-):
-    """Reserve slots from a data-dependent counter index."""
-    tid = wp.tid()
-    bin = bins[tid]
-    slot = wp.atomic_add(counters, bin, 1)
-    output[bin, slot] = values[tid]
-
-
-@wp.kernel
-def sliced_counter_kernel(
-    values: wp.array[wp.int32],
-    bins: wp.array[wp.int32],
-    counters: wp.array2d[wp.int32],
-    output: wp.array2d[wp.int32],
-):
-    """Reserve slots through a sliced counter view such as ``counters[bin]``."""
-    tid = wp.tid()
-    bin = bins[tid]
-    slot = wp.atomic_add(counters[bin], 0, 1)
-    output[bin, slot] = values[tid]
-
-
-@wp.kernel(module="unique", module_options={"deterministic": "run_to_run", "deterministic_max_records": 4})
-def loop_indexed_counter_kernel(
-    counts: wp.array[wp.int32],
-    bins: wp.array[wp.int32],
-    counters: wp.array[wp.int32],
-    output: wp.array2d[wp.int32],
-):
-    """Emit a data-dependent number of counter records to dynamic destinations."""
-    tid = wp.tid()
-    count = counts[tid]
-
-    for i in range(count):
-        bin = (bins[tid] + i) % 4
-        slot = wp.atomic_add(counters, bin, 1)
-        output[bin, slot] = tid * 10 + i
-
-
-@wp.kernel
-def counter_side_effect_kernel(
-    counter: wp.array[wp.int32],
-    output: wp.array[wp.float32],
-    scratch: wp.array[wp.float32],
-):
-    """Counter kernel with a normal array write that must not execute in phase 0."""
-    tid = wp.tid()
-    slot = wp.atomic_add(counter, 0, 1)
-    scratch[tid] = scratch[tid] + 1.0
-    output[slot] = float(tid)
-
-
-@wp.kernel(module="unique", module_options={"deterministic": "run_to_run"})
-def local_scratch_counter_kernel(
-    counter: wp.array[wp.int32],
-    output: wp.array[wp.int32],
-):
-    """Local scratch stores must execute in phase 0 when they control counters."""
-    tid = wp.tid()
-    scratch = wp.zeros(shape=(2,), dtype=wp.int32)
-    _det_set_int_scratch(scratch, 0, 1)
-    _det_set_int_scratch(scratch, 1, 1)
-
-    for i in range(2):
-        if scratch[i] != 0:
-            slot = wp.atomic_add(counter, 0, 1)
-            output[slot] = tid * 2 + i
-
-
-@wp.kernel(module="unique", module_options={"deterministic": "run_to_run"})
-def counter_with_atomic_xor_kernel(
-    counter: wp.array[wp.int32],
-    flag: wp.array[wp.int32],
-    output: wp.array[wp.int32],
-):
-    """Counter kernel with an unintercepted ``atomic_xor`` side effect."""
-    tid = wp.tid()
-    wp.atomic_xor(flag, 0, 1)
-    slot = wp.atomic_add(counter, 0, 1)
-    output[slot] = tid
-
-
-@wp.func
-def _det_set_flag(flag: wp.array[wp.int32], mask: wp.int32):
-    wp.atomic_xor(flag, 0, mask)
-
-
-@wp.kernel(module="unique", module_options={"deterministic": "run_to_run"})
-def counter_with_helper_atomic_xor_kernel(
-    counter: wp.array[wp.int32],
-    flag: wp.array[wp.int32],
-    output: wp.array[wp.int32],
-):
-    """Counter kernel that delegates the bitwise atomic to a ``@wp.func`` helper."""
-    tid = wp.tid()
-    _det_set_flag(flag, 1)
-    slot = wp.atomic_add(counter, 0, 1)
-    output[slot] = tid
-
-
-@wp.kernel(module="unique", module_options={"deterministic": "run_to_run"})
-def counter_with_consumed_atomic_xor_kernel(
-    flag: wp.array[wp.int32],
-    counter: wp.array[wp.int32],
-    output: wp.array[wp.int32],
-):
-    """Counter kernel whose control flow consumes a wrapped bitwise atomic's return."""
-    tid = wp.tid()
-    f = wp.atomic_xor(flag, 0, 1)
-    if f != 0:
-        slot = wp.atomic_add(counter, 0, 1)
-        output[slot] = tid
-
-
-@wp.kernel(module="unique", module_options={"deterministic": "run_to_run"})
-def counter_with_component_store_kernel(
-    counter: wp.array[wp.int32],
-    values: wp.array[wp.vec3],
-    output: wp.array[wp.int32],
-):
-    """Counter kernel that writes through the array-slot fast path."""
-    tid = wp.tid()
-    values[0].x = values[0].x + 1.0
-    slot = wp.atomic_add(counter, 0, 1)
-    output[slot] = tid
-
-
-@wp.func
-def _det_custom_replay_counter(counter: wp.array[wp.int32], tids: wp.array[wp.int32], tid: int):
-    slot = wp.atomic_add(counter, 0, 1)
-    tids[tid] = slot
-    return slot
-
-
-@wp.func_replay(_det_custom_replay_counter)
-def _replay_det_custom_replay_counter(counter: wp.array[wp.int32], tids: wp.array[wp.int32], tid: int):
-    return tids[tid]
-
-
-@wp.kernel(module="unique", module_options={"deterministic": "run_to_run"})
-def custom_replay_counter_kernel(
-    data: wp.array[wp.float32],
-    counter: wp.array[wp.int32],
-    tids: wp.array[wp.int32],
-    output: wp.array[wp.float32],
-):
-    tid = wp.tid()
-    slot = _det_custom_replay_counter(counter, tids, tid)
-    output[slot] = data[tid] * data[tid]
-
-
-# ---------------------------------------------------------------------------
-# Mixed kernels: counter plus scatter/reduce atomics.
-# ---------------------------------------------------------------------------
-
-
-@wp.kernel
-def mixed_pattern_kernel(
-    data: wp.array[wp.float32],
-    counter: wp.array[wp.int32],
-    output: wp.array[wp.float32],
-    accum: wp.array[wp.float32],
-):
-    """Counter allocation + accumulation in the same kernel."""
-    tid = wp.tid()
-    val = data[tid]
-
-    # Allocate a deterministic slot from a consumed-return atomic.
-    slot = wp.atomic_add(counter, 0, 1)
-    output[slot] = val
-
-    # Accumulate through the scatter/reduce path.
-    wp.atomic_add(accum, tid % 8, val)
-
-
-@wp.kernel
-def mixed_counter_int_atomic_kernel(
-    counter: wp.array[wp.int32],
-    output: wp.array[wp.int32],
-    accum: wp.array[wp.int32],
-):
-    """Integer side-effect atomic must not run during counter phase 0."""
-    tid = wp.tid()
-    wp.atomic_add(accum, tid % 8, 1)
-    slot = wp.atomic_add(counter, 0, 1)
-    output[slot] = tid
-
-
-# ---------------------------------------------------------------------------
-# Integer atomic kernels (should pass through unchanged when return unused)
-# ---------------------------------------------------------------------------
-
-
-@wp.kernel
-def int_atomic_add_kernel(
-    dest_indices: wp.array[wp.int32],
-    output: wp.array[wp.int32],
-):
-    """Integer atomic add (should be deterministic without transformation)."""
-    tid = wp.tid()
-    idx = dest_indices[tid]
-    wp.atomic_add(output, idx, 1)
-
-
-@wp.func
-def _det_lookup_value(values: wp.array[wp.float32], index: int) -> wp.float32:
-    return values[index]
-
-
-@wp.func_grad(_det_lookup_value)
-def _adj_det_lookup_value(values: wp.array[wp.float32], index: int, adj_ret: wp.float32):
-    wp.adjoint[values][index] += adj_ret
-
-
-@wp.kernel
-def custom_adjoint_lookup_kernel(
-    values: wp.array[wp.float32],
-    indices: wp.array[wp.int32],
-    output: wp.array[wp.float32],
-):
-    """Use a custom adjoint that accumulates into ``wp.adjoint[values]``."""
-    tid = wp.tid()
-    value = _det_lookup_value(values, indices[tid])
-    wp.atomic_add(output, 0, value)
-
-
-@wp.kernel
-def custom_adjoint_gather_kernel(
-    values: wp.array[wp.float32],
-    indices: wp.array[wp.int32],
-    output: wp.array[wp.float32],
-):
-    """Use a custom adjoint that scatters per-lane output gradients."""
-    tid = wp.tid()
-    output[tid] = _det_lookup_value(values, indices[tid])
-
-
-@wp.func
-def _det_custom_square_store(
-    i: int,
-    values: wp.array[wp.float32],
-    output: wp.array[wp.float32],
-    scratch: wp.array[wp.float32],
-):
-    output[i] = values[i] * values[i]
-
-
-@wp.func_grad(_det_custom_square_store)
-def _adj_det_custom_square_store(
-    i: int,
-    values: wp.array[wp.float32],
-    output: wp.array[wp.float32],
-    scratch: wp.array[wp.float32],
-):
-    scratch[i] = 0.0
-    wp.adjoint[values][i] += 2.0 * values[i] * wp.adjoint[output][i]
-
-
-@wp.kernel
-def custom_adjoint_store_kernel(
-    values: wp.array[wp.float32],
-    output: wp.array[wp.float32],
-    scratch: wp.array[wp.float32],
-):
-    tid = wp.tid()
-    _det_custom_square_store(tid, values, output, scratch)
-
-
-@wp.func
-def _not_guaranteed_lookup_value(values: wp.array[wp.float32], index: int) -> wp.float32:
-    return values[index]
-
-
-@wp.func_grad(_not_guaranteed_lookup_value)
-def _adj_not_guaranteed_lookup_value(values: wp.array[wp.float32], index: int, adj_ret: wp.float32):
-    wp.adjoint[values][index] += adj_ret
-
-
-@wp.kernel
-def custom_adjoint_not_guaranteed_kernel(
-    values: wp.array[wp.float32],
-    indices: wp.array[wp.int32],
-    output: wp.array[wp.float32],
-):
-    """Use a custom adjoint while deterministic mode is disabled."""
-    tid = wp.tid()
-    output[tid] = _not_guaranteed_lookup_value(values, indices[tid])
-
-
-# ---------------------------------------------------------------------------
-# Test functions
-# ---------------------------------------------------------------------------
-
-
-def test_scatter_add_reproducibility(test, device):
-    """Verify that float atomic_add produces bit-exact identical results across runs."""
-    if device.is_cpu:
-        test.skipTest("CPU execution is already deterministic")
-
-    n = 4096
-    out_size = 64
-    rng = np.random.default_rng(42)
-
-    data_np = rng.random(n, dtype=np.float32)
-    indices_np = rng.integers(0, out_size, size=n, dtype=np.int32)
-
-    data = wp.array(data_np, dtype=wp.float32, device=device)
-    indices = wp.array(indices_np, dtype=wp.int32, device=device)
-
-    results = []
-    for _ in range(10):
-        output = wp.zeros(out_size, dtype=wp.float32, device=device)
-        wp.launch(
-            scatter_add_kernel,
-            dim=n,
-            inputs=[data, indices],
-            outputs=[output],
-            device=device,
-        )
-        results.append(output.numpy().copy())
-
-    # All runs must produce bit-exact identical results.
-    for i in range(1, len(results)):
-        np.testing.assert_array_equal(
-            results[0],
-            results[i],
-            err_msg=f"Run 0 vs run {i} differ (deterministic mode should be bit-exact)",
-        )
-
-
-def test_gpu_to_gpu_mode_reproducibility(test, device):
-    """Verify the module-level ``gpu_to_gpu`` mode produces reproducible results."""
-    if device.is_cpu:
-        test.skipTest("CPU execution is already deterministic")
-
-    n = 1024
-    out_size = 16
-    rng = np.random.default_rng(44)
-
-    data_np = rng.random(n, dtype=np.float32)
-    indices_np = rng.integers(0, out_size, size=n, dtype=np.int32)
-
-    data = wp.array(data_np, dtype=wp.float32, device=device)
-    indices = wp.array(indices_np, dtype=wp.int32, device=device)
-
-    old_det = _get_test_module_options()["deterministic"]
-    try:
-        _set_test_module_options({"deterministic": "gpu_to_gpu"})
-        results = []
-        for _ in range(3):
-            output = wp.zeros(out_size, dtype=wp.float32, device=device)
-            wp.launch(scatter_add_kernel, dim=n, inputs=[data, indices], outputs=[output], device=device)
-            results.append(output.numpy().copy())
-        for i in range(1, len(results)):
-            np.testing.assert_array_equal(results[0], results[i])
-    finally:
-        _set_test_module_options({"deterministic": old_det})
-
-
-def test_gpu_to_gpu_matches_canonical_float32_reference(test, device):
-    """Verify ``gpu_to_gpu`` matches the canonical float32 CPU reduction order."""
-    if device.is_cpu:
-        test.skipTest("CPU execution is already deterministic")
-
-    data_np = np.array(
-        [
-            1.0e20,
-            1.0,
-            -1.0e20,
-            3.5,
-            -2.25,
-            2.0**-20,
-            1.0e10,
-            -1.0e10,
-            7.0,
-            -7.0,
-            9.0,
-            1.0e-7,
-        ],
-        dtype=np.float32,
-    )
-    indices_np = np.array([0, 0, 0, 1, 1, 1, 0, 0, 1, 1, 2, 2], dtype=np.int32)
-    out_size = 3
-
-    expected = _reference_scatter_add_float32(data_np, indices_np, out_size)
-
-    data = wp.array(data_np, dtype=wp.float32, device=device)
-    indices = wp.array(indices_np, dtype=wp.int32, device=device)
-
-    old_det = _get_test_module_options()["deterministic"]
-    try:
-        _set_test_module_options({"deterministic": "gpu_to_gpu"})
-        output = wp.zeros(out_size, dtype=wp.float32, device=device)
-        wp.launch(scatter_add_kernel, dim=data_np.shape[0], inputs=[data, indices], outputs=[output], device=device)
-        result = output.numpy()
-    finally:
-        _set_test_module_options({"deterministic": old_det})
-
-    np.testing.assert_array_equal(result.view(np.uint32), expected.view(np.uint32))
-
-
-def test_augassign_add_reproducibility(test, device):
-    """Verify += syntax (desugars to atomic_add) is also deterministic."""
-    if device.is_cpu:
-        test.skipTest("CPU execution is already deterministic")
-
-    n = 2048
-    out_size = 32
-    rng = np.random.default_rng(123)
-
-    data_np = rng.random(n, dtype=np.float32)
-    indices_np = rng.integers(0, out_size, size=n, dtype=np.int32)
-
-    data = wp.array(data_np, dtype=wp.float32, device=device)
-    indices = wp.array(indices_np, dtype=wp.int32, device=device)
-
-    results = []
-    for _ in range(10):
-        output = wp.zeros(out_size, dtype=wp.float32, device=device)
-        wp.launch(
-            augassign_add_kernel,
-            dim=n,
-            inputs=[data, indices],
-            outputs=[output],
-            device=device,
-        )
-        results.append(output.numpy().copy())
-
-    for i in range(1, len(results)):
-        np.testing.assert_array_equal(results[0], results[i])
-
-
-def test_scatter_add_correctness(test, device):
-    """Compare deterministic GPU results against CPU sequential execution."""
-    n = 2048
-    out_size = 32
-    rng = np.random.default_rng(99)
-
-    data_np = rng.random(n, dtype=np.float32)
-    indices_np = rng.integers(0, out_size, size=n, dtype=np.int32)
-
-    # CPU sequential reference (guaranteed deterministic).
-    expected = np.zeros(out_size, dtype=np.float32)
-    for i in range(n):
-        expected[indices_np[i]] += data_np[i]
-
-    data = wp.array(data_np, dtype=wp.float32, device=device)
-    indices = wp.array(indices_np, dtype=wp.int32, device=device)
-    output = wp.zeros(out_size, dtype=wp.float32, device=device)
-
-    wp.launch(
-        scatter_add_kernel,
-        dim=n,
-        inputs=[data, indices],
-        outputs=[output],
-        device=device,
-    )
-
-    result = output.numpy()
-    # Deterministic sum order may differ from Python loop order, so exact
-    # match is not guaranteed.  Check within reasonable tolerance.
-    np.testing.assert_allclose(result, expected, rtol=1e-4, atol=1e-5)
-
-
-def test_multi_array_atomic(test, device):
-    """Verify deterministic mode works with multiple target arrays."""
-    if device.is_cpu:
-        test.skipTest("CPU execution is already deterministic")
-
-    n = 1024
-    out_size = 16
-    rng = np.random.default_rng(77)
-
-    data_np = rng.random(n, dtype=np.float32)
-    indices_np = rng.integers(0, out_size, size=n, dtype=np.int32)
-
-    data = wp.array(data_np, dtype=wp.float32, device=device)
-    indices = wp.array(indices_np, dtype=wp.int32, device=device)
-
-    results_a, results_b, results_c = [], [], []
-    for _ in range(5):
-        out_a = wp.zeros(out_size, dtype=wp.float32, device=device)
-        out_b = wp.zeros(out_size, dtype=wp.float32, device=device)
-        out_c = wp.zeros(out_size, dtype=wp.float32, device=device)
-        wp.launch(
-            multi_array_atomic_kernel,
-            dim=n,
-            inputs=[data, indices],
-            outputs=[out_a, out_b, out_c],
-            device=device,
-        )
-        results_a.append(out_a.numpy().copy())
-        results_b.append(out_b.numpy().copy())
-        results_c.append(out_c.numpy().copy())
-
-    for i in range(1, len(results_a)):
-        np.testing.assert_array_equal(results_a[0], results_a[i])
-        np.testing.assert_array_equal(results_b[0], results_b[i])
-        np.testing.assert_array_equal(results_c[0], results_c[i])
-
-
-def test_atomic_sub_deterministic(test, device):
-    """Verify atomic_sub is deterministic."""
-    if device.is_cpu:
-        test.skipTest("CPU execution is already deterministic")
-
-    n = 2048
-    out_size = 32
-    rng = np.random.default_rng(55)
-
-    data_np = rng.random(n, dtype=np.float32)
-    indices_np = rng.integers(0, out_size, size=n, dtype=np.int32)
-
-    data = wp.array(data_np, dtype=wp.float32, device=device)
-    indices = wp.array(indices_np, dtype=wp.int32, device=device)
-
-    results = []
-    for _ in range(5):
-        output = wp.zeros(out_size, dtype=wp.float32, device=device)
-        wp.launch(
-            atomic_sub_kernel,
-            dim=n,
-            inputs=[data, indices],
-            outputs=[output],
-            device=device,
-        )
-        results.append(output.numpy().copy())
-
-    for i in range(1, len(results)):
-        np.testing.assert_array_equal(results[0], results[i])
-
-
-def test_atomic_add_2d(test, device):
-    """Verify deterministic mode with 2D array indexing."""
-    if device.is_cpu:
-        test.skipTest("CPU execution is already deterministic")
-
-    n = 1024
-    rows, cols = 8, 8
-    rng = np.random.default_rng(88)
-
-    data_np = rng.random(n, dtype=np.float32)
-    row_np = rng.integers(0, rows, size=n, dtype=np.int32)
-    col_np = rng.integers(0, cols, size=n, dtype=np.int32)
-
-    data = wp.array(data_np, dtype=wp.float32, device=device)
-    row_idx = wp.array(row_np, dtype=wp.int32, device=device)
-    col_idx = wp.array(col_np, dtype=wp.int32, device=device)
-
-    results = []
-    for _ in range(5):
-        output = wp.zeros(shape=(rows, cols), dtype=wp.float32, device=device)
-        wp.launch(
-            atomic_add_2d_kernel,
-            dim=n,
-            inputs=[data, row_idx, col_idx],
-            outputs=[output],
-            device=device,
-        )
-        results.append(output.numpy().copy())
-
-    for i in range(1, len(results)):
-        np.testing.assert_array_equal(results[0], results[i])
-
-
-def test_sliced_2d_array_atomic_add(test, device):
-    """Verify deterministic atomics through a sliced ``arr[row]`` view."""
-    if device.is_cpu:
-        test.skipTest("CPU execution is already deterministic")
-
-    n = 2048
-    rows, cols = 16, 16
-    rng = np.random.default_rng(101)
-
-    data_np = rng.random(n, dtype=np.float32)
-    row_np = rng.integers(0, rows, size=n, dtype=np.int32)
-    col_np = rng.integers(0, cols, size=n, dtype=np.int32)
-
-    data = wp.array(data_np, dtype=wp.float32, device=device)
-    row_idx = wp.array(row_np, dtype=wp.int32, device=device)
-    col_idx = wp.array(col_np, dtype=wp.int32, device=device)
-
-    expected = np.zeros((rows, cols), dtype=np.float32)
-    for i in range(n):
-        expected[row_np[i], col_np[i]] = np.float32(expected[row_np[i], col_np[i]] + data_np[i])
-
-    results = []
-    for _ in range(5):
-        output = wp.zeros(shape=(rows, cols), dtype=wp.float32, device=device)
-        wp.launch(
-            sliced_2d_atomic_add_kernel,
-            dim=n,
-            inputs=[data, row_idx, col_idx],
-            outputs=[output],
-            device=device,
-        )
-        results.append(output.numpy().copy())
-
-    for result in results:
-        np.testing.assert_allclose(result, expected, rtol=1e-5, atol=1e-5)
-    for i in range(1, len(results)):
-        np.testing.assert_array_equal(results[0], results[i])
-
-
-def test_sliced_3d_array_atomic_add(test, device):
-    """Verify deterministic atomics through a sliced ``arr[row, col]`` view."""
-    if device.is_cpu:
-        test.skipTest("CPU execution is already deterministic")
-
-    n = 2048
-    rows, cols, depth = 8, 8, 8
-    rng = np.random.default_rng(102)
-
-    data_np = rng.random(n, dtype=np.float32)
-    row_np = rng.integers(0, rows, size=n, dtype=np.int32)
-    col_np = rng.integers(0, cols, size=n, dtype=np.int32)
-    depth_np = rng.integers(0, depth, size=n, dtype=np.int32)
-
-    data = wp.array(data_np, dtype=wp.float32, device=device)
-    row_idx = wp.array(row_np, dtype=wp.int32, device=device)
-    col_idx = wp.array(col_np, dtype=wp.int32, device=device)
-    depth_idx = wp.array(depth_np, dtype=wp.int32, device=device)
-
-    expected = np.zeros((rows, cols, depth), dtype=np.float32)
-    for i in range(n):
-        expected[row_np[i], col_np[i], depth_np[i]] = np.float32(
-            expected[row_np[i], col_np[i], depth_np[i]] + data_np[i]
-        )
-
-    results = []
-    for _ in range(5):
-        output = wp.zeros(shape=(rows, cols, depth), dtype=wp.float32, device=device)
-        wp.launch(
-            sliced_3d_atomic_add_kernel,
-            dim=n,
-            inputs=[data, row_idx, col_idx, depth_idx],
-            outputs=[output],
-            device=device,
-        )
-        results.append(output.numpy().copy())
-
-    for result in results:
-        np.testing.assert_allclose(result, expected, rtol=1e-5, atol=1e-5)
-    for i in range(1, len(results)):
-        np.testing.assert_array_equal(results[0], results[i])
-
-
-def test_strided_1d_view_atomic_add(test, device):
-    """Atomic add through a non-contiguous ``base[::2]`` view: odd slots untouched."""
-    if device.is_cpu:
-        test.skipTest("CPU execution is already deterministic")
-
-    n = 1024
-    base_size = 128
-    view_size = base_size // 2
-    rng = np.random.default_rng(303)
-
-    data_np = rng.random(n, dtype=np.float32)
-    dest_np = rng.integers(0, view_size, size=n, dtype=np.int32)
-
-    expected_view = np.zeros(view_size, dtype=np.float32)
-    for i in range(n):
-        expected_view[dest_np[i]] = np.float32(expected_view[dest_np[i]] + data_np[i])
-
-    sentinel = np.float32(-7777.0)
-
-    data = wp.array(data_np, dtype=wp.float32, device=device)
-    dest = wp.array(dest_np, dtype=wp.int32, device=device)
-
-    results = []
-    for _ in range(3):
-        base = wp.full(shape=(base_size,), value=sentinel, dtype=wp.float32, device=device)
-        view = base[::2]
-        test.assertFalse(view.is_contiguous)
-        wp.launch(scatter_add_kernel, dim=n, inputs=[data, dest], outputs=[view], device=device)
-        results.append(base.numpy().copy())
-
-    for result in results:
-        even = result[0::2]
-        odd = result[1::2]
-        np.testing.assert_allclose(even, sentinel + expected_view, rtol=1e-5, atol=1e-5)
-        np.testing.assert_array_equal(odd, np.full(view_size, sentinel, dtype=np.float32))
-    for i in range(1, len(results)):
-        np.testing.assert_array_equal(results[0], results[i])
-
-
-def test_zero_stride_view_atomic_add(test, device):
-    """Atomic add through a zero-stride view reduces to the one physical slot."""
-    if device.is_cpu:
-        test.skipTest("CPU execution is already deterministic")
-
-    n = 1024
-    logical_size = 8
-    rng = np.random.default_rng(306)
-
-    data_np = rng.random(n, dtype=np.float32)
-    dest_np = rng.integers(0, logical_size, size=n, dtype=np.int32)
-
-    sentinel = np.float32(-11.0)
-    expected = sentinel
-    for value in data_np:
-        expected = np.float32(expected + value)
-
-    data = wp.array(data_np, dtype=wp.float32, device=device)
-    dest = wp.array(dest_np, dtype=wp.int32, device=device)
-
-    results = []
-    for _ in range(3):
-        base = wp.full(shape=(1,), value=sentinel, dtype=wp.float32, device=device)
-        view = wp.array(
-            ptr=base.ptr,
-            dtype=wp.float32,
-            shape=(logical_size,),
-            strides=(0,),
-            capacity=4,
-            device=device,
-        )
-        test.assertFalse(view.is_contiguous)
-        wp.launch(scatter_add_kernel, dim=n, inputs=[data, dest], outputs=[view], device=device)
-        results.append(base.numpy().copy())
-
-    for result in results:
-        np.testing.assert_allclose(result, np.array([expected], dtype=np.float32), rtol=1e-5, atol=1e-5)
-    for i in range(1, len(results)):
-        np.testing.assert_array_equal(results[0], results[i])
-
-
-def test_column_slice_atomic_add(test, device):
-    """Atomic add through a non-contiguous ``base[:, col]`` view: other columns untouched."""
-    if device.is_cpu:
-        test.skipTest("CPU execution is already deterministic")
-
-    n = 1024
-    rows, cols = 64, 4
-    target_col = 2
-    rng = np.random.default_rng(304)
-
-    data_np = rng.random(n, dtype=np.float32)
-    dest_np = rng.integers(0, rows, size=n, dtype=np.int32)
-
-    expected_col = np.zeros(rows, dtype=np.float32)
-    for i in range(n):
-        expected_col[dest_np[i]] = np.float32(expected_col[dest_np[i]] + data_np[i])
-
-    sentinel = np.float32(-3333.0)
-
-    data = wp.array(data_np, dtype=wp.float32, device=device)
-    dest = wp.array(dest_np, dtype=wp.int32, device=device)
-
-    results = []
-    for _ in range(3):
-        base = wp.full(shape=(rows, cols), value=sentinel, dtype=wp.float32, device=device)
-        view = base[:, target_col]
-        test.assertFalse(view.is_contiguous)
-        wp.launch(scatter_add_kernel, dim=n, inputs=[data, dest], outputs=[view], device=device)
-        results.append(base.numpy().copy())
-
-    for result in results:
-        np.testing.assert_allclose(result[:, target_col], sentinel + expected_col, rtol=1e-5, atol=1e-5)
-        other_cols = np.delete(result, target_col, axis=1)
-        np.testing.assert_array_equal(other_cols, np.full(other_cols.shape, sentinel, dtype=np.float32))
-    for i in range(1, len(results)):
-        np.testing.assert_array_equal(results[0], results[i])
-
-
-def test_transposed_2d_atomic_add(test, device):
-    """Atomic add through a transposed 2D view: ``view[i, j]`` writes to ``base[j, i]``."""
-    if device.is_cpu:
-        test.skipTest("CPU execution is already deterministic")
-
-    n = 2048
-    base_rows, base_cols = 16, 8
-    view_rows, view_cols = base_cols, base_rows
-    rng = np.random.default_rng(305)
-
-    data_np = rng.random(n, dtype=np.float32)
-    row_np = rng.integers(0, view_rows, size=n, dtype=np.int32)
-    col_np = rng.integers(0, view_cols, size=n, dtype=np.int32)
-
-    expected_view = np.zeros((view_rows, view_cols), dtype=np.float32)
-    for i in range(n):
-        expected_view[row_np[i], col_np[i]] = np.float32(expected_view[row_np[i], col_np[i]] + data_np[i])
-
-    data = wp.array(data_np, dtype=wp.float32, device=device)
-    row_idx = wp.array(row_np, dtype=wp.int32, device=device)
-    col_idx = wp.array(col_np, dtype=wp.int32, device=device)
-
-    results = []
-    for _ in range(3):
-        base = wp.zeros(shape=(base_rows, base_cols), dtype=wp.float32, device=device)
-        view = base.transpose([1, 0])
-        test.assertFalse(view.is_contiguous)
-        wp.launch(
-            atomic_add_2d_kernel,
-            dim=n,
-            inputs=[data, row_idx, col_idx],
-            outputs=[view],
-            device=device,
-        )
-        results.append(base.numpy().copy())
-
-    for result in results:
-        np.testing.assert_allclose(result, expected_view.T, rtol=1e-5, atol=1e-5)
-    for i in range(1, len(results)):
-        np.testing.assert_array_equal(results[0], results[i])
-
-
-def test_atomic_half_deterministic(test, device):
-    """Verify deterministic mode with float16 atomics."""
-    if device.is_cpu:
-        test.skipTest("CPU execution is already deterministic")
-
-    n = 1024
-    out_size = 16
-    rng = np.random.default_rng(78)
-
-    data_np = rng.random(n, dtype=np.float32).astype(np.float16)
-    indices_np = rng.integers(0, out_size, size=n, dtype=np.int32)
-
-    data = wp.array(data_np, dtype=wp.float16, device=device)
-    indices = wp.array(indices_np, dtype=wp.int32, device=device)
-
-    results = []
-    for _ in range(5):
-        output = wp.zeros(out_size, dtype=wp.float16, device=device)
-        wp.launch(
-            atomic_half_kernel,
-            dim=n,
-            inputs=[data, indices],
-            outputs=[output],
-            device=device,
-        )
-        results.append(output.numpy().copy())
-
-    for i in range(1, len(results)):
-        np.testing.assert_array_equal(results[0], results[i])
-
-
-def test_atomic_bfloat16_deterministic(test, device):
-    """Verify deterministic mode with bfloat16 atomics."""
-    if device.is_cpu:
-        test.skipTest("CPU execution is already deterministic")
-    if device.arch < 80:
-        test.skipTest("bfloat16 atomics require CUDA architecture >= 80")
-
-    n = 64
-    out_size = 8
-    rng = np.random.default_rng(91)
-
-    # Small integers are exactly representable in bfloat16, so correctness does
-    # not depend on emulating bfloat16 rounding in the host reference.
-    data_np = rng.integers(1, 5, size=n).astype(np.float32)
-    indices_np = rng.integers(0, out_size, size=n, dtype=np.int32)
-
-    expected_add = np.zeros(out_size, dtype=np.float32)
-    expected_min = np.full(out_size, 16.0, dtype=np.float32)
-    expected_max = np.zeros(out_size, dtype=np.float32)
-    for value, index in zip(data_np, indices_np, strict=True):
-        expected_add[index] += value
-        expected_min[index] = min(expected_min[index], value)
-        expected_max[index] = max(expected_max[index], value)
-
-    data = wp.array(data_np, dtype=wp.bfloat16, device=device)
-    indices = wp.array(indices_np, dtype=wp.int32, device=device)
-
-    add_bits = []
-    for _ in range(3):
-        output = wp.zeros(out_size, dtype=wp.bfloat16, device=device)
-        wp.launch(
-            atomic_bfloat16_kernel,
-            dim=n,
-            inputs=[data, indices],
-            outputs=[output],
-            device=device,
-        )
-        add_bits.append(_bfloat16_numpy_bits(output.numpy()))
-        np.testing.assert_allclose(_bfloat16_numpy_to_float32(output.numpy()), expected_add, rtol=0.0, atol=0.0)
-
-    test.assertTrue(
-        any(target.scalar_dtype is wp.bfloat16 for target in atomic_bfloat16_kernel.adj.det_meta.scatter_targets)
-    )
-    for i in range(1, len(add_bits)):
-        np.testing.assert_array_equal(add_bits[0], add_bits[i])
-
-    min_bits = []
-    max_bits = []
-    for _ in range(3):
-        output_min = wp.full(out_size, value=wp.bfloat16(16.0), dtype=wp.bfloat16, device=device)
-        output_max = wp.zeros(out_size, dtype=wp.bfloat16, device=device)
-        wp.launch(
-            atomic_bfloat16_minmax_kernel,
-            dim=n,
-            inputs=[data, indices],
-            outputs=[output_min, output_max],
-            device=device,
-        )
-        min_bits.append(_bfloat16_numpy_bits(output_min.numpy()))
-        max_bits.append(_bfloat16_numpy_bits(output_max.numpy()))
-        np.testing.assert_allclose(_bfloat16_numpy_to_float32(output_min.numpy()), expected_min, rtol=0.0, atol=0.0)
-        np.testing.assert_allclose(_bfloat16_numpy_to_float32(output_max.numpy()), expected_max, rtol=0.0, atol=0.0)
-
-    test.assertTrue(
-        any(target.scalar_dtype is wp.bfloat16 for target in atomic_bfloat16_minmax_kernel.adj.det_meta.scatter_targets)
-    )
-    for i in range(1, len(min_bits)):
-        np.testing.assert_array_equal(min_bits[0], min_bits[i])
-        np.testing.assert_array_equal(max_bits[0], max_bits[i])
-
-
-def test_atomic_double_deterministic(test, device):
-    """Verify deterministic mode with float64 atomics."""
-    if device.is_cpu:
-        test.skipTest("CPU execution is already deterministic")
-
-    n = 2048
-    out_size = 32
-    rng = np.random.default_rng(66)
-
-    data_np = rng.random(n).astype(np.float64)
-    indices_np = rng.integers(0, out_size, size=n, dtype=np.int32)
-
-    data = wp.array(data_np, dtype=wp.float64, device=device)
-    indices = wp.array(indices_np, dtype=wp.int32, device=device)
-
-    results = []
-    for _ in range(5):
-        output = wp.zeros(out_size, dtype=wp.float64, device=device)
-        wp.launch(
-            atomic_double_kernel,
-            dim=n,
-            inputs=[data, indices],
-            outputs=[output],
-            device=device,
-        )
-        results.append(output.numpy().copy())
-
-    for i in range(1, len(results)):
-        np.testing.assert_array_equal(results[0], results[i])
-
-
-def test_vec3_atomic_add_deterministic(test, device):
-    """Verify deterministic mode for composite ``wp.vec3`` atomic adds."""
-    if device.is_cpu:
-        test.skipTest("CPU execution is already deterministic")
-
-    n = 1024
-    out_size = 16
-    rng = np.random.default_rng(67)
-
-    data_np = rng.standard_normal((n, 3), dtype=np.float32)
-    indices_np = rng.integers(0, out_size, size=n, dtype=np.int32)
-
-    data = wp.array(data_np, dtype=wp.vec3, device=device)
-    indices = wp.array(indices_np, dtype=wp.int32, device=device)
-
-    results = []
-    for _ in range(5):
-        output = wp.zeros(out_size, dtype=wp.vec3, device=device)
-        wp.launch(vec3_scatter_add_kernel, dim=n, inputs=[data, indices], outputs=[output], device=device)
-        results.append(output.numpy().copy())
-
-    expected = np.zeros((out_size, 3), dtype=np.float32)
-    for i in range(n):
-        expected[indices_np[i]] += data_np[i]
-
-    for result in results:
-        np.testing.assert_allclose(result, expected, rtol=1e-5, atol=1e-5)
-    for i in range(1, len(results)):
-        np.testing.assert_array_equal(results[0], results[i])
-
-
-def test_vec3_atomic_minmax_deterministic(test, device):
-    """Verify deterministic component-wise ``wp.vec3`` min/max reductions."""
-    if device.is_cpu:
-        test.skipTest("CPU execution is already deterministic")
-
-    n = 2048
-    rng = np.random.default_rng(68)
-    points_np = rng.standard_normal((n, 3), dtype=np.float32)
-    points = wp.array(points_np, dtype=wp.vec3, device=device)
-
-    mins = []
-    maxs = []
-    for _ in range(5):
-        out_min = wp.empty(1, dtype=wp.vec3, device=device)
-        out_max = wp.empty(1, dtype=wp.vec3, device=device)
-        out_min.fill_(wp.vec3(np.inf, np.inf, np.inf))
-        out_max.fill_(wp.vec3(-np.inf, -np.inf, -np.inf))
-        wp.launch(vec3_atomic_minmax_kernel, dim=n, inputs=[points], outputs=[out_min, out_max], device=device)
-        mins.append(out_min.numpy().copy())
-        maxs.append(out_max.numpy().copy())
-
-    expected_min = np.min(points_np, axis=0, keepdims=True)
-    expected_max = np.max(points_np, axis=0, keepdims=True)
-
-    for result in mins:
-        np.testing.assert_allclose(result, expected_min, rtol=0.0, atol=0.0)
-    for result in maxs:
-        np.testing.assert_allclose(result, expected_max, rtol=0.0, atol=0.0)
-    for i in range(1, len(mins)):
-        np.testing.assert_array_equal(mins[0], mins[i])
-        np.testing.assert_array_equal(maxs[0], maxs[i])
-
-
-def test_mat33_atomic_add_deterministic(test, device):
-    """Verify deterministic mode for composite ``wp.mat33`` atomic adds."""
-    if device.is_cpu:
-        test.skipTest("CPU execution is already deterministic")
-
-    n = 512
-    out_size = 8
-    rng = np.random.default_rng(69)
-
-    data_np = rng.standard_normal((n, 3, 3), dtype=np.float32)
-    indices_np = rng.integers(0, out_size, size=n, dtype=np.int32)
-
-    data = wp.array(data_np, dtype=wp.mat33, device=device)
-    indices = wp.array(indices_np, dtype=wp.int32, device=device)
-
-    results = []
-    for _ in range(5):
-        output = wp.zeros(out_size, dtype=wp.mat33, device=device)
-        wp.launch(mat33_scatter_add_kernel, dim=n, inputs=[data, indices], outputs=[output], device=device)
-        results.append(output.numpy().copy())
-
-    expected = np.zeros((out_size, 3, 3), dtype=np.float32)
-    for i in range(n):
-        expected[indices_np[i]] += data_np[i]
-
-    for result in results:
-        np.testing.assert_allclose(result, expected, rtol=1e-5, atol=1e-5)
-    for i in range(1, len(results)):
-        np.testing.assert_array_equal(results[0], results[i])
-
-
-def test_triple_scatter_capacity_estimate(test, device):
-    """Verify kernels with >2 scatters per thread do not overflow the buffer."""
-    if device.is_cpu:
-        test.skipTest("CPU execution is already deterministic")
-
-    n = 512
-    rng = np.random.default_rng(12)
-    data_np = rng.random(n, dtype=np.float32)
-    data = wp.array(data_np, dtype=wp.float32, device=device)
-
-    results = []
-    for _ in range(3):
-        output = wp.zeros(1, dtype=wp.float32, device=device)
-        wp.launch(triple_scatter_add_kernel, dim=n, inputs=[data], outputs=[output], device=device)
-        results.append(output.numpy().copy())
-
-    expected = np.array([6.0 * data_np.sum()], dtype=np.float32)
-    for result in results:
-        np.testing.assert_allclose(result, expected, rtol=1e-5, atol=1e-5)
-    for i in range(1, len(results)):
-        np.testing.assert_array_equal(results[0], results[i])
-
-
-def test_loop_scatter_max_records_override(test, device):
-    """Verify ``deterministic_max_records`` handles dynamic loop emission counts."""
-    if device.is_cpu:
-        test.skipTest("CPU execution is already deterministic")
-
-    n = 256
-    rng = np.random.default_rng(71)
-    data_np = rng.random(n, dtype=np.float32)
-    counts_np = np.full(n, 4, dtype=np.int32)
-
-    data = wp.array(data_np, dtype=wp.float32, device=device)
-    counts = wp.array(counts_np, dtype=wp.int32, device=device)
-
-    expected = np.array([np.dot(data_np, counts_np).astype(np.float32)], dtype=np.float32)
-
-    results = []
-    for _ in range(3):
-        output = wp.zeros(1, dtype=wp.float32, device=device)
-        wp.launch(loop_scatter_add_kernel, dim=n, inputs=[data, counts], outputs=[output], device=device)
-        results.append(output.numpy().copy())
-
-    for result in results:
-        np.testing.assert_allclose(result, expected, rtol=1e-5, atol=1e-5)
-    for i in range(1, len(results)):
-        np.testing.assert_array_equal(results[0], results[i])
-
-
-def test_scatter_overflow_reports_error(test, device):
-    """Verify an underprovisioned dynamic scatter reports overflow to the host."""
-    if device.is_cpu:
-        test.skipTest("CPU execution is already deterministic")
-
-    n = 2048
-    data = wp.ones(n, dtype=wp.float32, device=device)
-    counts = wp.full(n, value=2, dtype=wp.int32, device=device)
-    output = wp.zeros(1, dtype=wp.float32, device=device)
-
-    with test.assertRaisesRegex(RuntimeError, "Deterministic scatter buffer overflow"):
-        wp.launch(underprovisioned_loop_scatter_kernel, dim=n, inputs=[data, counts], outputs=[output], device=device)
-
-
-def test_scatter_large_launch_rejected(test, device):
-    """Verify scatter key packing fails clearly before oversized launches."""
-    if device.is_cpu:
-        test.skipTest("CPU execution is already deterministic")
-
-    data = wp.ones(1, dtype=wp.float32, device=device)
-    indices = wp.zeros(1, dtype=wp.int32, device=device)
-    output = wp.zeros(1, dtype=wp.float32, device=device)
-
-    with test.assertRaisesRegex(RuntimeError, "up to 2\\^32 threads"):
-        wp.launch(
-            scatter_add_kernel,
-            dim=(65536, 65537),
-            inputs=[data, indices],
-            outputs=[output],
-            device=device,
-        )
-
-
-def test_mixed_reduce_ops_same_array(test, device):
-    """Verify mixed reduction families on one array are rejected in deterministic mode."""
-    if device.is_cpu:
-        test.skipTest("CPU execution is already deterministic")
-
-    data_np = np.full(4, 0.05, dtype=np.float32)
-    data = wp.array(data_np, dtype=wp.float32, device=device)
-    output = wp.zeros(1, dtype=wp.float32, device=device)
-
-    with test.assertRaisesRegex(Exception, "does not support mixing"):
-
-        @wp.kernel(module="unique", module_options={"deterministic": "run_to_run"})
-        def mixed_reduce_op_same_array_local_kernel(
-            data: wp.array[wp.float32],
-            output: wp.array[wp.float32],
-        ):
-            tid = wp.tid()
-            wp.atomic_add(output, 0, data[tid])
-            wp.atomic_max(output, 0, 1.0)
-
-        wp.launch(
-            mixed_reduce_op_same_array_local_kernel,
-            dim=data_np.shape[0],
-            inputs=[data],
-            outputs=[output],
-            device=device,
-        )
-
-
-def test_struct_field_counter_atomic(test, device):
-    """Verify deterministic counters work when the target array lives in a struct field."""
-    if device.is_cpu:
-        test.skipTest("CPU execution is already deterministic")
-
-    rng = np.random.default_rng(34)
-    data_np = rng.random(64, dtype=np.float32)
-    counts_np = rng.integers(0, 4, size=64, dtype=np.int32)
-    expected = []
-    for tid, count in enumerate(counts_np):
-        for i in range(int(count)):
-            expected.append(np.float32(data_np[tid] + np.float32(i) * np.float32(0.5)))
-    expected = np.asarray(expected, dtype=np.float32)
-
-    data = wp.array(data_np, dtype=wp.float32, device=device)
-    counts = wp.array(counts_np, dtype=wp.int32, device=device)
-
-    results = []
-    counter_values = []
-    for _ in range(3):
-        writer = _DetStructCounterWriter()
-        writer.counter = wp.zeros(1, dtype=wp.int32, device=device)
-        writer.output = wp.zeros(expected.shape[0] + 4, dtype=wp.float32, device=device)
-        wp.launch(struct_field_counter_kernel, dim=64, inputs=[data, counts, writer], device=device)
-        counter_values.append(int(writer.counter.numpy()[0]))
-        results.append(writer.output.numpy()[: expected.shape[0]].copy())
-
-    np.testing.assert_array_equal(np.array(counter_values), np.full(3, expected.shape[0], dtype=np.int32))
-    for i in range(1, len(results)):
-        np.testing.assert_array_equal(results[0], results[i])
-    np.testing.assert_array_equal(results[0].view(np.uint32), expected.view(np.uint32))
-
-
-def test_struct_field_helper_counter_atomic(test, device):
-    """Verify helper-function counters work when the target array lives in a struct field."""
-    if device.is_cpu:
-        test.skipTest("CPU execution is already deterministic")
-
-    rng = np.random.default_rng(35)
-    data_np = rng.random(64, dtype=np.float32)
-    flags_np = (rng.random(64) > 0.4).astype(np.int32)
-    expected = data_np[flags_np != 0]
-
-    data = wp.array(data_np, dtype=wp.float32, device=device)
-    flags = wp.array(flags_np, dtype=wp.int32, device=device)
-
-    results = []
-    counter_values = []
-    for _ in range(3):
-        writer = _DetStructCounterWriter()
-        writer.counter = wp.zeros(1, dtype=wp.int32, device=device)
-        writer.output = wp.zeros(expected.shape[0] + 4, dtype=wp.float32, device=device)
-        wp.launch(struct_field_helper_counter_kernel, dim=64, inputs=[data, flags, writer], device=device)
-        counter_values.append(int(writer.counter.numpy()[0]))
-        results.append(writer.output.numpy()[: expected.shape[0]].copy())
-
-    np.testing.assert_array_equal(np.array(counter_values), np.full(3, expected.shape[0], dtype=np.int32))
-    for i in range(1, len(results)):
-        np.testing.assert_array_equal(results[0], results[i])
-    np.testing.assert_array_equal(results[0].view(np.uint32), expected.view(np.uint32))
-
-
-def test_metadata_refresh_struct_tile_stack(test, device):
-    """Verify deterministic cache metadata refresh handles struct tile-stack dtypes."""
-    del device
-    options = struct_tile_stack_kernel.module.resolve_options(wp.config) | {"output_arch": None}
-    struct_tile_stack_kernel.adj.build(None, options)
-    test.assertIsNotNone(struct_tile_stack_kernel.adj.det_meta)
-
-
-def test_config_deterministic_max_records_default(test, device):
-    """Verify new modules inherit ``wp.config.deterministic_max_records``."""
-    del device
-    old_max_records = wp.config.deterministic_max_records
-    try:
-        wp.config.deterministic_max_records = 7
-
-        @wp.kernel(module="unique")
-        def _config_max_records_kernel(output: wp.array[wp.float32]):
-            tid = wp.tid()
-            wp.atomic_add(output, tid % 2, 1.0)
-
-        options = _config_max_records_kernel.module.resolve_options(wp.config)
-        test.assertEqual(options["deterministic_max_records"], 7)
-    finally:
-        wp.config.deterministic_max_records = old_max_records
-
-
-def test_helper_name_collision(test, device):
-    """Verify deterministic helpers stay unique for labels with the same sanitized form."""
-    if device.is_cpu:
-        test.skipTest("CPU execution is already deterministic")
-
-    data_np = np.linspace(0.25, 2.0, 32, dtype=np.float32)
-    data = wp.array(data_np, dtype=wp.float32, device=device)
-    direct = wp.zeros(1, dtype=wp.float32, device=device)
-    field = wp.zeros(1, dtype=wp.float32, device=device)
-    holder = _DetNameCollisionStruct()
-    holder.b = field
-
-    wp.launch(helper_name_collision_kernel, dim=data_np.shape[0], inputs=[direct, holder, data], device=device)
-
-    np.testing.assert_allclose(direct.numpy(), np.array([data_np.sum()], dtype=np.float32), rtol=1e-5, atol=1e-5)
-    np.testing.assert_allclose(field.numpy(), np.array([2.0 * data_np.sum()], dtype=np.float32), rtol=1e-5, atol=1e-5)
-
-
-def test_counter_reproducibility(test, device):
-    """Verify consumed-return counters produce deterministic slot assignments."""
-    if device.is_cpu:
-        test.skipTest("CPU execution is already deterministic")
-
-    n = 1024
-    rng = np.random.default_rng(33)
-    data_np = rng.random(n, dtype=np.float32)
-    data = wp.array(data_np, dtype=wp.float32, device=device)
-
-    results = []
-    for _ in range(10):
-        counter = wp.zeros(1, dtype=wp.int32, device=device)
-        output = wp.zeros(n, dtype=wp.float32, device=device)
-        wp.launch(
-            counter_kernel,
-            dim=n,
-            inputs=[data, counter],
-            outputs=[output],
-            device=device,
-        )
-        results.append(output.numpy().copy())
-
-    for i in range(1, len(results)):
-        np.testing.assert_array_equal(
-            results[0],
-            results[i],
-            err_msg=f"Counter run 0 vs run {i} differ",
-        )
-
-
-def test_counter_phase0_suppresses_array_writes(test, device):
-    """Verify non-counter array stores are skipped during the counting pass."""
-    if device.is_cpu:
-        test.skipTest("CPU execution is already deterministic")
-
-    n = 128
-    counter = wp.zeros(1, dtype=wp.int32, device=device)
-    output = wp.zeros(n, dtype=wp.float32, device=device)
-    scratch = wp.zeros(n, dtype=wp.float32, device=device)
-
-    wp.launch(counter_side_effect_kernel, dim=n, inputs=[counter], outputs=[output, scratch], device=device)
-
-    np.testing.assert_array_equal(scratch.numpy(), np.ones(n, dtype=np.float32))
-    test.assertEqual(int(counter.numpy()[0]), n)
-
-
-def test_counter_phase0_suppresses_array_writes_before_helper(test, device):
-    """Verify phase 0 suppresses stores before helper functions with counters."""
-    if device.is_cpu:
-        test.skipTest("CPU execution is already deterministic")
-
-    n = 128
-    counter = wp.zeros(1, dtype=wp.int32, device=device)
-    output = wp.zeros(n, dtype=wp.float32, device=device)
-    scratch = wp.zeros(n, dtype=wp.float32, device=device)
-
-    wp.launch(helper_counter_side_effect_kernel, dim=n, inputs=[counter], outputs=[output, scratch], device=device)
-
-    np.testing.assert_array_equal(scratch.numpy(), np.ones(n, dtype=np.float32))
-    test.assertEqual(int(counter.numpy()[0]), n)
-
-
-def test_counter_phase0_suppresses_helper_array_writes(test, device):
-    """Verify phase 0 suppresses pure write helpers called from counter kernels."""
-    if device.is_cpu:
-        test.skipTest("CPU execution is already deterministic")
-
-    n = 128
-    counter = wp.zeros(1, dtype=wp.int32, device=device)
-    output = wp.zeros(n, dtype=wp.float32, device=device)
-    scratch = wp.zeros(n, dtype=wp.float32, device=device)
-
-    wp.launch(counter_with_helper_store_kernel, dim=n, inputs=[counter], outputs=[output, scratch], device=device)
-
-    np.testing.assert_array_equal(scratch.numpy(), np.ones(n, dtype=np.float32))
-    test.assertEqual(int(counter.numpy()[0]), n)
-
-
-def test_counter_phase0_preserves_local_scratch_writes(test, device):
-    """Verify phase 0 executes local scratch stores that affect counter calls."""
-    if device.is_cpu:
-        test.skipTest("CPU execution is already deterministic")
-
-    n = 64
-    counter = wp.zeros(1, dtype=wp.int32, device=device)
-    output = wp.full(2 * n, value=-1, dtype=wp.int32, device=device)
-
-    wp.launch(local_scratch_counter_kernel, dim=n, inputs=[counter], outputs=[output], device=device)
-
-    test.assertEqual(int(counter.numpy()[0]), 2 * n)
-    np.testing.assert_array_equal(output.numpy(), np.arange(2 * n, dtype=np.int32))
-
-
-def test_counter_phase0_skips_unintercepted_atomic_xor(test, device):
-    """Verify ``atomic_xor`` in a counter kernel fires once, not twice."""
-    if device.is_cpu:
-        test.skipTest("CPU execution is already deterministic")
-
-    counter = wp.zeros(1, dtype=wp.int32, device=device)
-    flag = wp.zeros(1, dtype=wp.int32, device=device)
-    output = wp.full(1, value=-1, dtype=wp.int32, device=device)
-
-    wp.launch(counter_with_atomic_xor_kernel, dim=1, inputs=[counter, flag], outputs=[output], device=device)
-
-    test.assertEqual(int(flag.numpy()[0]), 1)
-    test.assertEqual(int(counter.numpy()[0]), 1)
-    test.assertEqual(int(output.numpy()[0]), 0)
-
-
-def test_counter_phase0_skips_atomic_xor_in_helper_func(test, device):
-    """Verify ``atomic_xor`` in a @wp.func called from a counter kernel fires once."""
-    if device.is_cpu:
-        test.skipTest("CPU execution is already deterministic")
-
-    counter = wp.zeros(1, dtype=wp.int32, device=device)
-    flag = wp.zeros(1, dtype=wp.int32, device=device)
-    output = wp.full(1, value=-1, dtype=wp.int32, device=device)
-
-    wp.launch(counter_with_helper_atomic_xor_kernel, dim=1, inputs=[counter, flag], outputs=[output], device=device)
-
-    test.assertEqual(int(flag.numpy()[0]), 1)
-    test.assertEqual(int(counter.numpy()[0]), 1)
-    test.assertEqual(int(output.numpy()[0]), 0)
-
-
-def test_counter_consumed_bitwise_atomic_rejected(test, device):
-    """Consuming a bitwise atomic return inside a two-pass body must be rejected."""
-    if device.is_cpu:
-        test.skipTest("CPU execution is already deterministic")
-
-    flag = wp.full(1, value=1, dtype=wp.int32, device=device)
-    counter = wp.zeros(1, dtype=wp.int32, device=device)
-    output = wp.full(1, value=-1, dtype=wp.int32, device=device)
-
-    with test.assertRaisesRegex(Exception, r"wp\.atomic_xor"):
-        wp.launch(
-            counter_with_consumed_atomic_xor_kernel,
-            dim=1,
-            inputs=[flag, counter],
-            outputs=[output],
-            device=device,
-        )
-
-
-def test_counter_phase0_suppresses_component_stores(test, device):
-    """Verify component-level array stores do not double-execute in counter mode."""
-    if device.is_cpu:
-        test.skipTest("CPU execution is already deterministic")
-
-    counter = wp.zeros(1, dtype=wp.int32, device=device)
-    values = wp.zeros(1, dtype=wp.vec3, device=device)
-    output = wp.full(1, value=-1, dtype=wp.int32, device=device)
-
-    wp.launch(counter_with_component_store_kernel, dim=1, inputs=[counter, values], outputs=[output], device=device)
-
-    np.testing.assert_array_equal(values.numpy(), np.array([[1.0, 0.0, 0.0]], dtype=np.float32))
-    test.assertEqual(int(counter.numpy()[0]), 1)
-    test.assertEqual(int(output.numpy()[0]), 0)
-
-
-def test_counter_correctness(test, device):
-    """Verify consumed-return counters write all data (no lost elements)."""
-    n = 512
-    rng = np.random.default_rng(44)
-    data_np = rng.random(n, dtype=np.float32)
-    data = wp.array(data_np, dtype=wp.float32, device=device)
-
-    counter = wp.zeros(1, dtype=wp.int32, device=device)
-    output = wp.zeros(n, dtype=wp.float32, device=device)
-    wp.launch(
-        counter_kernel,
-        dim=n,
-        inputs=[data, counter],
-        outputs=[output],
-        device=device,
-    )
-
-    # Counter should equal n.
-    count = int(counter.numpy()[0])
-    test.assertEqual(count, n)
-
-    # Output should contain a permutation of data.
-    result = sorted(output.numpy().tolist())
-    expected = sorted(data_np.tolist())
-    np.testing.assert_allclose(result, expected, rtol=1e-6)
-
-
-def test_counter_nonzero_initial_value(test, device):
-    """Verify consumed-return counters preserve a non-zero initial value.
-
-    Before the fix to ``make_counter_prefix_states_kernel`` the deterministic
-    prefix scan started from zero, ignoring the counter's existing value.  In
-    real multi-launch workloads (e.g. particle emission across timesteps) this
-    silently overwrote slots from earlier launches and produced a wrong final
-    count.
-    """
-    if device.is_cpu:
-        test.skipTest("CPU execution is already deterministic")
-
-    n = 64
-    k = 20  # non-zero initial counter value
-
-    rng = np.random.default_rng(777)
-    data_np = rng.random(n, dtype=np.float32)
-    data = wp.array(data_np, dtype=wp.float32, device=device)
-
-    counter = wp.full(1, value=k, dtype=wp.int32, device=device)
-    sentinel = np.float32(-7.0)
-    output = wp.full(k + n, value=sentinel, dtype=wp.float32, device=device)
-
-    wp.launch(counter_kernel, dim=n, inputs=[data, counter], outputs=[output], device=device)
-
-    # Counter must reflect the seed plus the n new allocations.
-    test.assertEqual(int(counter.numpy()[0]), k + n)
-
-    # Slots [0, K) must be untouched.
-    np.testing.assert_array_equal(
-        output.numpy()[:k],
-        np.full(k, sentinel, dtype=np.float32),
-        err_msg="Initial slots were overwritten -- counter scan ignored its seed.",
-    )
-
-    # Slots [K, K+n) must contain a permutation of the input data.
-    written = sorted(output.numpy()[k:].tolist())
-    expected = sorted(data_np.tolist())
-    np.testing.assert_allclose(written, expected, rtol=1e-6)
-
-
-def test_counter_multi_launch_accumulates(test, device):
-    """Verify consumed-return counters accumulate correctly across launches.
-
-    Realistic multi-launch use case (e.g. emitting particles over several
-    timesteps): each launch must append its allocations on top of the previous
-    launches' slots, not overwrite them.
-    """
-    if device.is_cpu:
-        test.skipTest("CPU execution is already deterministic")
-
-    n = 32
-    num_launches = 3
-    capacity = n * num_launches
-
-    rng = np.random.default_rng(778)
-    counter = wp.zeros(1, dtype=wp.int32, device=device)
-    output = wp.full(capacity, value=np.float32(-1.0), dtype=wp.float32, device=device)
-
-    all_data = []
-    for _ in range(num_launches):
-        data_np = rng.random(n, dtype=np.float32)
-        all_data.append(data_np)
-        data = wp.array(data_np, dtype=wp.float32, device=device)
-        wp.launch(counter_kernel, dim=n, inputs=[data, counter], outputs=[output], device=device)
-
-    test.assertEqual(int(counter.numpy()[0]), capacity)
-
-    # All n*num_launches values must be present in the output, none overwritten.
-    written = sorted(output.numpy().tolist())
-    expected = sorted(np.concatenate(all_data).tolist())
-    np.testing.assert_allclose(written, expected, rtol=1e-6)
-
-
-def test_counter_variable_total_writeback(test, device):
-    """Verify deterministic counters publish variable total counts on device."""
-    if device.is_cpu:
-        test.skipTest("CUDA deterministic counter writeback required")
-
-    counts_np = np.array([2, 0, 3, 1, 0, 4, 1, 0], dtype=np.int32)
-    expected = []
-    for tid, count in enumerate(counts_np):
-        expected.extend([tid * 10 + i for i in range(count)])
-    expected_np = np.array(expected, dtype=np.int32)
-
-    counts = wp.array(counts_np, dtype=wp.int32, device=device)
-    counter = wp.zeros(1, dtype=wp.int32, device=device)
-    output = wp.full(expected_np.shape[0], value=-1, dtype=wp.int32, device=device)
-
-    wp.launch(
-        variable_counter_kernel, dim=counts_np.shape[0], inputs=[counts, counter], outputs=[output], device=device
-    )
-
-    np.testing.assert_array_equal(counter.numpy(), np.array([expected_np.shape[0]], dtype=np.int32))
-    np.testing.assert_array_equal(output.numpy(), expected_np)
-
-
-def test_counter_nonzero_index(test, device):
-    """Verify consumed-return counters support fixed nonzero indices."""
-    if device.is_cpu:
-        test.skipTest("CPU execution is already deterministic")
-
-    n = 32
-    results = []
-    counters = []
-
-    for _ in range(3):
-        counter = wp.zeros(2, dtype=wp.int32, device=device)
-        output = wp.full(n, value=-1, dtype=wp.int32, device=device)
-        wp.launch(static_index_counter_kernel, dim=n, inputs=[counter], outputs=[output], device=device)
-        counters.append(counter.numpy().copy())
-        results.append(output.numpy().copy())
-
-    expected_counter = np.array([0, n], dtype=np.int32)
-    expected_output = np.arange(n, dtype=np.int32)
-    for counter in counters:
-        np.testing.assert_array_equal(counter, expected_counter)
-    for result in results:
-        np.testing.assert_array_equal(result, expected_output)
-
-
-def test_counter_indexed_destinations(test, device):
-    """Verify consumed-return counters support data-dependent counter indices."""
-    if device.is_cpu:
-        test.skipTest("CPU execution is already deterministic")
-
-    n = 128
-    bin_count = 5
-    rng = np.random.default_rng(92)
-    bins_np = rng.integers(0, bin_count, size=n, dtype=np.int32)
-    values_np = np.arange(n, dtype=np.int32) + 100
-
-    expected_counts = np.bincount(bins_np, minlength=bin_count).astype(np.int32)
-    expected_output = np.full((bin_count, n), -1, dtype=np.int32)
-    offsets = np.zeros(bin_count, dtype=np.int32)
-    for tid, bin in enumerate(bins_np):
-        slot = offsets[bin]
-        expected_output[bin, slot] = values_np[tid]
-        offsets[bin] += 1
-
-    values = wp.array(values_np, dtype=wp.int32, device=device)
-    bins = wp.array(bins_np, dtype=wp.int32, device=device)
-
-    results = []
-    counters = []
-    for _ in range(3):
-        counter = wp.zeros(bin_count, dtype=wp.int32, device=device)
-        output = wp.full((bin_count, n), value=-1, dtype=wp.int32, device=device)
-        wp.launch(indexed_counter_kernel, dim=n, inputs=[values, bins, counter], outputs=[output], device=device)
-        counters.append(counter.numpy().copy())
-        results.append(output.numpy().copy())
-
-    for counter in counters:
-        np.testing.assert_array_equal(counter, expected_counts)
-    for result in results:
-        np.testing.assert_array_equal(result, expected_output)
-
-
-def test_counter_sliced_destinations(test, device):
-    """Verify consumed-return counters support sliced counter-array views."""
-    if device.is_cpu:
-        test.skipTest("CPU execution is already deterministic")
-
-    n = 96
-    bin_count = 4
-    bins_np = (np.arange(n, dtype=np.int32) * 3) % bin_count
-    values_np = np.arange(n, dtype=np.int32) + 1000
-
-    expected_counts = np.bincount(bins_np, minlength=bin_count).astype(np.int32).reshape(bin_count, 1)
-    expected_output = np.full((bin_count, n), -1, dtype=np.int32)
-    offsets = np.zeros(bin_count, dtype=np.int32)
-    for tid, bin in enumerate(bins_np):
-        slot = offsets[bin]
-        expected_output[bin, slot] = values_np[tid]
-        offsets[bin] += 1
-
-    values = wp.array(values_np, dtype=wp.int32, device=device)
-    bins = wp.array(bins_np, dtype=wp.int32, device=device)
-
-    counters = wp.zeros((bin_count, 1), dtype=wp.int32, device=device)
-    output = wp.full((bin_count, n), value=-1, dtype=wp.int32, device=device)
-    wp.launch(sliced_counter_kernel, dim=n, inputs=[values, bins, counters], outputs=[output], device=device)
-
-    np.testing.assert_array_equal(counters.numpy(), expected_counts)
-    np.testing.assert_array_equal(output.numpy(), expected_output)
-
-
-def test_counter_indexed_dynamic_loop(test, device):
-    """Verify dynamic loops can reserve multiple records for dynamic counter indices."""
-    if device.is_cpu:
-        test.skipTest("CPU execution is already deterministic")
-
-    n = 64
-    bin_count = 4
-    counts_np = (np.arange(n, dtype=np.int32) % 4).astype(np.int32)
-    bins_np = ((np.arange(n, dtype=np.int32) * 5) % bin_count).astype(np.int32)
-
-    expected_counts = np.zeros(bin_count, dtype=np.int32)
-    expected_output = np.full((bin_count, n * 4), -1, dtype=np.int32)
-    for tid in range(n):
-        for i in range(counts_np[tid]):
-            bin = (bins_np[tid] + i) % bin_count
-            slot = expected_counts[bin]
-            expected_output[bin, slot] = tid * 10 + i
-            expected_counts[bin] += 1
-
-    counts = wp.array(counts_np, dtype=wp.int32, device=device)
-    bins = wp.array(bins_np, dtype=wp.int32, device=device)
-    counters = wp.zeros(bin_count, dtype=wp.int32, device=device)
-    output = wp.full((bin_count, n * 4), value=-1, dtype=wp.int32, device=device)
-
-    wp.launch(loop_indexed_counter_kernel, dim=n, inputs=[counts, bins, counters], outputs=[output], device=device)
-
-    np.testing.assert_array_equal(counters.numpy(), expected_counts)
-    np.testing.assert_array_equal(output.numpy(), expected_output)
-
-
-def test_counter_int64_rejected(test, device):
-    """Verify unsupported non-int32 consumed-return counters fail clearly."""
-    if device.is_cpu:
-        test.skipTest("CPU execution is already deterministic")
-
-    with test.assertRaisesRegex(Exception, "int32 counter arrays"):
-
-        @wp.kernel(module="unique", module_options={"deterministic": "run_to_run"})
-        def counter_int64_kernel(
-            counter: wp.array[wp.int64],
-            output: wp.array[wp.float32],
-        ):
-            """Unsupported non-int32 consumed-return counter."""
-            tid = wp.tid()
-            slot = wp.atomic_add(counter, 0, wp.int64(1))
-            output[slot] = float(tid)
-
-        counter = wp.zeros(1, dtype=wp.int64, device=device)
-        output = wp.zeros(8, dtype=wp.float32, device=device)
-        wp.launch(counter_int64_kernel, dim=8, inputs=[counter], outputs=[output], device=device)
-
-
-def test_counter_non_add_atomic_rejected(test, device):
-    """Verify consumed-return deterministic counters only accept ``atomic_add``."""
-    if device.is_cpu:
-        test.skipTest("CPU execution is already deterministic")
-
-    with test.assertRaisesRegex(Exception, "only for atomic_add"):
-
-        @wp.kernel(module="unique", module_options={"deterministic": "run_to_run"})
-        def counter_sub_kernel(
-            counter: wp.array[wp.int32],
-            output: wp.array[wp.float32],
-        ):
-            """Unsupported consumed-return counter using ``atomic_sub``."""
-            tid = wp.tid()
-            slot = wp.atomic_sub(counter, 0, 1)
-            output[slot] = float(tid)
-
-        counter = wp.zeros(1, dtype=wp.int32, device=device)
-        output = wp.zeros(8, dtype=wp.float32, device=device)
-        wp.launch(counter_sub_kernel, dim=8, inputs=[counter], outputs=[output], device=device)
-
-    with test.assertRaisesRegex(Exception, "only for atomic_add"):
-
-        @wp.kernel(module="unique", module_options={"deterministic": "run_to_run"})
-        def counter_max_kernel(
-            counter: wp.array[wp.int32],
-            output: wp.array[wp.float32],
-        ):
-            """Unsupported consumed-return counter using ``atomic_max``."""
-            tid = wp.tid()
-            old = wp.atomic_max(counter, 0, tid)
-            output[tid] = float(old)
-
-        counter = wp.zeros(1, dtype=wp.int32, device=device)
-        output = wp.zeros(8, dtype=wp.float32, device=device)
-        wp.launch(counter_max_kernel, dim=8, inputs=[counter], outputs=[output], device=device)
-
-
-def test_float_atomic_consumed_return_rejected(test, device):
-    """Verify float atomics with consumed-not-assigned returns fail clearly."""
-    if device.is_cpu:
-        test.skipTest("CPU execution is already deterministic")
-
-    pattern = "consumed-return counter atomics only for int32 counter arrays"
-
-    with test.assertRaisesRegex(Exception, pattern):
-
-        @wp.kernel(module="unique", module_options={"deterministic": "run_to_run"})
-        def float_leader_if_kernel(
-            total: wp.array[wp.float32],
-            log: wp.array[wp.int32],
-        ):
-            """Return consumed by an ``if`` test."""
-            tid = wp.tid()
-            if wp.atomic_add(total, 0, 1.0) == 0.0:
-                log[0] = tid
-
-        total = wp.zeros(1, dtype=wp.float32, device=device)
-        log = wp.zeros(1, dtype=wp.int32, device=device)
-        wp.launch(float_leader_if_kernel, dim=8, inputs=[total], outputs=[log], device=device)
-
-    with test.assertRaisesRegex(Exception, pattern):
-
-        @wp.kernel(module="unique", module_options={"deterministic": "run_to_run"})
-        def float_nested_call_in_if_kernel(
-            total: wp.array[wp.float32],
-            out: wp.array[wp.float32],
-        ):
-            """Return consumed by a nested call inside an ``if`` test."""
-            tid = wp.tid()
-            if wp.abs(wp.atomic_add(total, 0, 1.0)) > 0.0:
-                out[tid] = 1.0
-
-        total = wp.zeros(1, dtype=wp.float32, device=device)
-        out = wp.zeros(8, dtype=wp.float32, device=device)
-        wp.launch(float_nested_call_in_if_kernel, dim=8, inputs=[total], outputs=[out], device=device)
-
-    with test.assertRaisesRegex(Exception, pattern):
-
-        @wp.kernel(module="unique", module_options={"deterministic": "run_to_run"})
-        def float_nested_bare_atomic_kernel(
-            total: wp.array[wp.float32],
-            other: wp.array[wp.float32],
-        ):
-            """Return consumed as the value argument of a bare atomic statement."""
-            wp.atomic_add(total, 0, wp.atomic_add(other, 0, 1.0))
-
-        total = wp.zeros(1, dtype=wp.float32, device=device)
-        other = wp.zeros(1, dtype=wp.float32, device=device)
-        wp.launch(float_nested_bare_atomic_kernel, dim=8, inputs=[total], outputs=[other], device=device)
-
-
-def test_float_atomic_bare_statement_allowed(test, device):
-    """Verify bare-statement float atomics still lower to the scatter path."""
-    if device.is_cpu:
-        test.skipTest("CPU execution is already deterministic")
-
-    @wp.kernel(module="unique", module_options={"deterministic": "run_to_run"})
-    def bare_float_accum_kernel(
-        contribs: wp.array[wp.float32],
-        accum: wp.array[wp.float32],
-    ):
-        tid = wp.tid()
-        wp.atomic_add(accum, tid % 4, contribs[tid])
-
-    n = 32
-    rng = np.random.default_rng(0)
-    contribs_np = rng.random(n, dtype=np.float32)
-    contribs = wp.array(contribs_np, dtype=wp.float32, device=device)
-    accum = wp.zeros(4, dtype=wp.float32, device=device)
-
-    wp.launch(bare_float_accum_kernel, dim=n, inputs=[contribs], outputs=[accum], device=device)
-
-    expected = np.zeros(4, dtype=np.float32)
-    for i in range(n):
-        expected[i % 4] += contribs_np[i]
-
-    np.testing.assert_allclose(accum.numpy(), expected, rtol=0, atol=1e-5)
-
-
-def test_scatter_capacity_overflow_rejected(test, device):
-    """Verify oversized deterministic scatter buffers fail before allocation."""
-    del device
-
-    meta = loop_scatter_add_kernel.adj.det_meta
-    test.assertIsNotNone(meta)
-    test.assertTrue(meta.has_scatter)
-
-    with test.assertRaisesRegex(RuntimeError, "int32 limit"):
-        wp_deterministic.allocate_scatter_buffers(
-            meta.scatter_targets,
-            meta,
-            wp_deterministic.DETERMINISTIC_SCATTER_MAX_CAPACITY + 1,
-            wp.get_device("cpu"),
-            max_records=1,
-        )
-
-
-def test_conditional_counter(test, device):
-    """Verify stream compaction with a conditional counter."""
-    if device.is_cpu:
-        test.skipTest("CPU execution is already deterministic")
-
-    n = 2048
-    threshold = 0.5
-    rng = np.random.default_rng(55)
-    data_np = rng.random(n, dtype=np.float32)
-    expected_count = int(np.sum(data_np > threshold))
-
-    data = wp.array(data_np, dtype=wp.float32, device=device)
-
-    results = []
-    counts = []
-    for _ in range(5):
-        counter = wp.zeros(1, dtype=wp.int32, device=device)
-        output = wp.zeros(n, dtype=wp.float32, device=device)
-        wp.launch(
-            conditional_counter_kernel,
-            dim=n,
-            inputs=[data, threshold, counter],
-            outputs=[output],
-            device=device,
-        )
-        counts.append(int(counter.numpy()[0]))
-        results.append(output.numpy()[:expected_count].copy())
-
-    # Count should be correct.
-    for c in counts:
-        test.assertEqual(c, expected_count)
-
-    # Results should be identical across runs.
-    for i in range(1, len(results)):
-        np.testing.assert_array_equal(results[0], results[i])
-
-
-def test_mixed_pattern(test, device):
-    """Verify kernel with both counter and accumulation atomics."""
-    if device.is_cpu:
-        test.skipTest("CPU execution is already deterministic")
-
-    n = 512
-    rng = np.random.default_rng(77)
-    data_np = rng.random(n, dtype=np.float32)
-    data = wp.array(data_np, dtype=wp.float32, device=device)
-
-    results_out, results_accum = [], []
-    for _ in range(5):
-        counter = wp.zeros(1, dtype=wp.int32, device=device)
-        output = wp.zeros(n, dtype=wp.float32, device=device)
-        accum = wp.zeros(8, dtype=wp.float32, device=device)
-        wp.launch(
-            mixed_pattern_kernel,
-            dim=n,
-            inputs=[data, counter],
-            outputs=[output, accum],
-            device=device,
-        )
-        results_out.append(output.numpy().copy())
-        results_accum.append(accum.numpy().copy())
-
-    for i in range(1, len(results_out)):
-        np.testing.assert_array_equal(results_out[0], results_out[i])
-        np.testing.assert_array_equal(results_accum[0], results_accum[i])
-
-
-def test_counter_with_integer_accumulation(test, device):
-    """Verify integer atomics in counter kernels do not run during phase 0."""
-    if device.is_cpu:
-        test.skipTest("CPU execution is already deterministic")
-
-    n = 512
-    counter = wp.zeros(1, dtype=wp.int32, device=device)
-    output = wp.zeros(n, dtype=wp.int32, device=device)
-    accum = wp.zeros(8, dtype=wp.int32, device=device)
-
-    wp.launch(mixed_counter_int_atomic_kernel, dim=n, inputs=[counter], outputs=[output, accum], device=device)
-
-    np.testing.assert_array_equal(counter.numpy(), [n])
-    np.testing.assert_array_equal(accum.numpy(), np.full(8, n // 8, dtype=np.int32))
-
-
-def test_int_atomic_passthrough(test, device):
-    """Verify integer atomics (return unused) work without overhead."""
-    n = 1024
-    out_size = 16
-    rng = np.random.default_rng(11)
-    indices_np = rng.integers(0, out_size, size=n, dtype=np.int32)
-    indices = wp.array(indices_np, dtype=wp.int32, device=device)
-
-    output = wp.zeros(out_size, dtype=wp.int32, device=device)
-    wp.launch(
-        int_atomic_add_kernel,
-        dim=n,
-        inputs=[indices],
-        outputs=[output],
-        device=device,
-    )
-
-    result = output.numpy()
-    # Verify correctness: count of each index.
-    expected = np.bincount(indices_np, minlength=out_size).astype(np.int32)
-    np.testing.assert_array_equal(result, expected)
-
-
-def test_module_option_override(test, device):
-    """Verify per-module deterministic option works."""
-
-    # Create a kernel with a per-module deterministic override.
-    @wp.kernel(module_options={"deterministic": "gpu_to_gpu"}, module="unique")
-    def per_kernel_det(
-        data: wp.array[wp.float32],
-        output: wp.array[wp.float32],
-    ):
-        tid = wp.tid()
-        wp.atomic_add(output, tid % 4, data[tid])
-
-    n = 256
-    rng = np.random.default_rng(22)
-    data_np = rng.random(n, dtype=np.float32)
-    data = wp.array(data_np, dtype=wp.float32, device=device)
-
-    # Ensure the shared module is disabled but the unique module option still works.
-    old_det = _get_test_module_options()["deterministic"]
-    try:
-        _set_test_module_options({"deterministic": "not_guaranteed"})
-        output = wp.zeros(4, dtype=wp.float32, device=device)
-        wp.launch(per_kernel_det, dim=n, inputs=[data], outputs=[output], device=device)
-        result = output.numpy()
-        # Basic sanity: sum should be approximately correct.
-        for bin_idx in range(4):
-            mask = np.arange(n) % 4 == bin_idx
-            expected_sum = data_np[mask].sum()
-            np.testing.assert_allclose(result[bin_idx], expected_sum, rtol=1e-4)
-    finally:
-        _set_test_module_options({"deterministic": old_det})
-
-
-def test_deterministic_closure_kernel(test, device):
-    """Verify deterministic closure kernels remain reproducible and distinct."""
-    if device.is_cpu:
-        test.skipTest("CPU execution is already deterministic")
-
-    kernel_a = _make_deterministic_closure_kernel(_det_closure_transform_a)
-    kernel_b = _make_deterministic_closure_kernel(_det_closure_transform_b)
-
-    test.assertIsNot(kernel_a, kernel_b)
-    test.assertNotEqual(kernel_a.module.name, kernel_b.module.name)
-
-    n = 512
-    rng = np.random.default_rng(30)
-    data_np = rng.random(n, dtype=np.float32)
-    data = wp.array(data_np, dtype=wp.float32, device=device)
-
-    results_a = []
-    results_b = []
-    for _ in range(3):
-        out_a = wp.zeros(8, dtype=wp.float32, device=device)
-        out_b = wp.zeros(8, dtype=wp.float32, device=device)
-        wp.launch(kernel_a, dim=n, inputs=[data], outputs=[out_a], device=device)
-        wp.launch(kernel_b, dim=n, inputs=[data], outputs=[out_b], device=device)
-        results_a.append(out_a.numpy().copy())
-        results_b.append(out_b.numpy().copy())
-
-    for i in range(1, len(results_a)):
-        np.testing.assert_array_equal(results_a[0], results_a[i])
-        np.testing.assert_array_equal(results_b[0], results_b[i])
-
-    test.assertFalse(np.array_equal(results_a[0], results_b[0]))
-
-
-def test_deterministic_func_kernel(test, device):
-    """Verify deterministic atomics inside ``@wp.func`` calls remain reproducible."""
-    if device.is_cpu:
-        test.skipTest("CPU execution is already deterministic")
-
-    n = 512
-    out_size = 16
-    rng = np.random.default_rng(74)
-    data_np = rng.random(n, dtype=np.float32)
-    indices_np = rng.integers(0, out_size, size=n, dtype=np.int32)
-
-    data = wp.array(data_np, dtype=wp.float32, device=device)
-    indices = wp.array(indices_np, dtype=wp.int32, device=device)
-    results = []
-    for _ in range(3):
-        output = wp.zeros(out_size, dtype=wp.float32, device=device)
-        wp.launch(func_scatter_add_kernel, dim=n, inputs=[data, indices], outputs=[output], device=device)
-        results.append(output.numpy().copy())
-
-    for result in results:
-        np.testing.assert_allclose(
-            result, _reference_scatter_add_float32(data_np, indices_np, out_size), rtol=1e-6, atol=1e-6
-        )
-    for i in range(1, len(results)):
-        np.testing.assert_array_equal(results[0], results[i])
-
-
-def test_nested_deterministic_func_kernel(test, device):
-    """Verify deterministic helper args propagate through nested ``@wp.func`` calls."""
-    if device.is_cpu:
-        test.skipTest("CPU execution is already deterministic")
-
-    n = 512
-    out_size = 16
-    rng = np.random.default_rng(75)
-    data_np = rng.random(n, dtype=np.float32)
-    indices_np = rng.integers(0, out_size, size=n, dtype=np.int32)
-
-    data = wp.array(data_np, dtype=wp.float32, device=device)
-    indices = wp.array(indices_np, dtype=wp.int32, device=device)
-    results = []
-    for _ in range(3):
-        accum = wp.zeros(out_size, dtype=wp.float32, device=device)
-        wp.launch(nested_func_scatter_add_kernel, dim=n, inputs=[data, indices], outputs=[accum], device=device)
-        results.append(accum.numpy().copy())
-
-    for result in results:
-        np.testing.assert_allclose(
-            result, _reference_scatter_add_float32(data_np, indices_np, out_size), rtol=1e-6, atol=1e-6
-        )
-    for i in range(1, len(results)):
-        np.testing.assert_array_equal(results[0], results[i])
-
-
-def test_record_cmd_deterministic_launch(test, device):
-    """Verify ``record_cmd=True`` works for deterministic CUDA launches."""
-    if device.is_cpu:
-        test.skipTest("CPU execution is already deterministic")
-
-    n = 128
-    out_size = 8
-    rng = np.random.default_rng(19)
-
-    data_np = rng.random(n, dtype=np.float32)
-    indices_np = rng.integers(0, out_size, size=n, dtype=np.int32)
-
-    data = wp.array(data_np, dtype=wp.float32, device=device)
-    indices = wp.array(indices_np, dtype=wp.int32, device=device)
-    output = wp.zeros(out_size, dtype=wp.float32, device=device)
-
-    cmd = wp.launch(
-        scatter_add_kernel,
-        dim=n,
-        inputs=[data, indices],
-        outputs=[output],
-        device=device,
-        record_cmd=True,
-    )
-
-    np.testing.assert_array_equal(output.numpy(), np.zeros(out_size, dtype=np.float32))
-
-    cmd.launch()
-    expected = output.numpy().copy()
-
-    output_2 = wp.zeros(out_size, dtype=wp.float32, device=device)
-    cmd.set_param_by_name("output", output_2)
-    cmd.launch()
-
-    np.testing.assert_array_equal(output_2.numpy(), expected)
-
-    output_3 = wp.zeros(out_size, dtype=wp.float32, device=device)
-    with test.assertRaisesRegex(RuntimeError, "raw ctypes"):
-        cmd.set_param_by_name_from_ctype("output", output_3.__ctype__())
-
-    cmd.set_param_by_name("output", output_3)
-    cmd.launch()
-
-    np.testing.assert_array_equal(output_3.numpy(), expected)
-
-
-def test_graph_capture_deterministic_launch(test, device):
-    """Verify deterministic scatter launches can be captured and replayed."""
-    if device.is_cpu:
-        test.skipTest("Graph capture requires CUDA")
-
-    n = 256
-    rng = np.random.default_rng(29)
-    data_np = rng.random(n, dtype=np.float32)
-    indices_np = rng.integers(0, 8, size=n, dtype=np.int32)
-
-    data = wp.array(data_np, dtype=wp.float32, device=device)
-    indices = wp.array(indices_np, dtype=wp.int32, device=device)
-    output = wp.zeros(8, dtype=wp.float32, device=device)
-
-    wp.launch(scatter_add_kernel, dim=n, inputs=[data, indices], outputs=[output], device=device)
-    output.zero_()
-
-    with wp.ScopedCapture(device, force_module_load=False) as capture:
-        wp.launch(scatter_add_kernel, dim=n, inputs=[data, indices], outputs=[output], device=device)
-
-    test.assertGreater(len(capture.graph._deterministic_buffer_refs), 0)
-    test.assertTrue(any(getattr(ref, "dtype", None) is wp.uint8 for ref in capture.graph._deterministic_buffer_refs))
-    gc.collect()
-
-    wp.capture_launch(capture.graph)
-    first = output.numpy().copy()
-
-    output.zero_()
-    wp.capture_launch(capture.graph)
-    second = output.numpy().copy()
-
-    np.testing.assert_array_equal(first, second)
-
-
-def test_graph_capture_sliced_array(test, device):
-    """Verify deterministic sliced-array atomics can be captured and replayed."""
-    if device.is_cpu:
-        test.skipTest("Graph capture requires CUDA")
-
-    # TestDeterministic.setUpClass enables run-to-run deterministic mode for
-    # every registered test in this module.
-    test.assertEqual(_get_test_module_options()["deterministic"], "run_to_run")
-
-    n = 256
-    rows, cols = 8, 8
-    rng = np.random.default_rng(201)
-
-    data_np = rng.random(n, dtype=np.float32)
-    row_np = rng.integers(0, rows, size=n, dtype=np.int32)
-    col_np = rng.integers(0, cols, size=n, dtype=np.int32)
-
-    data = wp.array(data_np, dtype=wp.float32, device=device)
-    row_idx = wp.array(row_np, dtype=wp.int32, device=device)
-    col_idx = wp.array(col_np, dtype=wp.int32, device=device)
-    output = wp.zeros(shape=(rows, cols), dtype=wp.float32, device=device)
-
-    wp.launch(sliced_2d_atomic_add_kernel, dim=n, inputs=[data, row_idx, col_idx], outputs=[output], device=device)
-    output.zero_()
-
-    with wp.ScopedCapture(device, force_module_load=False) as capture:
-        wp.launch(sliced_2d_atomic_add_kernel, dim=n, inputs=[data, row_idx, col_idx], outputs=[output], device=device)
-
-    wp.capture_launch(capture.graph)
-    first = output.numpy().copy()
-
-    output.zero_()
-    wp.capture_launch(capture.graph)
-    second = output.numpy().copy()
-
-    np.testing.assert_array_equal(first, second)
-
-
-def test_graph_capture_deterministic_closure_kernel(test, device):
-    """Verify deterministic closure kernels can be captured and replayed."""
-    if device.is_cpu:
-        test.skipTest("Graph capture requires CUDA")
-
-    kernel = _make_deterministic_closure_kernel(_det_closure_transform_a)
-
-    n = 256
-    rng = np.random.default_rng(31)
-    data_np = rng.random(n, dtype=np.float32)
-    data = wp.array(data_np, dtype=wp.float32, device=device)
-    output = wp.zeros(8, dtype=wp.float32, device=device)
-
-    wp.launch(kernel, dim=n, inputs=[data], outputs=[output], device=device)
-    output.zero_()
-
-    with wp.ScopedCapture(device, force_module_load=False) as capture:
-        wp.launch(kernel, dim=n, inputs=[data], outputs=[output], device=device)
-
-    wp.capture_launch(capture.graph)
-    first = output.numpy().copy()
-
-    output.zero_()
-    wp.capture_launch(capture.graph)
-    second = output.numpy().copy()
-
-    np.testing.assert_array_equal(first, second)
-
-
-def test_graph_capture_deterministic_func_kernel(test, device):
-    """Verify deterministic ``@wp.func`` atomics remain capture-safe."""
-    if device.is_cpu:
-        test.skipTest("Graph capture requires CUDA")
-
-    n = 256
-    rng = np.random.default_rng(76)
-    data_np = rng.random(n, dtype=np.float32)
-    indices_np = rng.integers(0, 8, size=n, dtype=np.int32)
-
-    data = wp.array(data_np, dtype=wp.float32, device=device)
-    indices = wp.array(indices_np, dtype=wp.int32, device=device)
-    output = wp.zeros(8, dtype=wp.float32, device=device)
-
-    wp.launch(func_scatter_add_kernel, dim=n, inputs=[data, indices], outputs=[output], device=device)
-    output.zero_()
-
-    with wp.ScopedCapture(device, force_module_load=False) as capture:
-        wp.launch(func_scatter_add_kernel, dim=n, inputs=[data, indices], outputs=[output], device=device)
-
-    wp.capture_launch(capture.graph)
-    first = output.numpy().copy()
-
-    output.zero_()
-    wp.capture_launch(capture.graph)
-    second = output.numpy().copy()
-
-    np.testing.assert_array_equal(first, second)
-
-
-def test_graph_capture_vec3_atomic_minmax(test, device):
-    """Verify composite deterministic reductions remain capture-safe."""
-    if device.is_cpu:
-        test.skipTest("Graph capture requires CUDA")
-
-    n = 512
-    rng = np.random.default_rng(70)
-    points_np = rng.standard_normal((n, 3), dtype=np.float32)
-    points = wp.array(points_np, dtype=wp.vec3, device=device)
-
-    out_min = wp.empty(1, dtype=wp.vec3, device=device)
-    out_max = wp.empty(1, dtype=wp.vec3, device=device)
-    min_init = wp.vec3(np.inf, np.inf, np.inf)
-    max_init = wp.vec3(-np.inf, -np.inf, -np.inf)
-
-    out_min.fill_(min_init)
-    out_max.fill_(max_init)
-    wp.launch(vec3_atomic_minmax_kernel, dim=n, inputs=[points], outputs=[out_min, out_max], device=device)
-    out_min.fill_(min_init)
-    out_max.fill_(max_init)
-
-    with wp.ScopedCapture(device, force_module_load=False) as capture:
-        wp.launch(vec3_atomic_minmax_kernel, dim=n, inputs=[points], outputs=[out_min, out_max], device=device)
-
-    wp.capture_launch(capture.graph)
-    first_min = out_min.numpy().copy()
-    first_max = out_max.numpy().copy()
-
-    out_min.fill_(min_init)
-    out_max.fill_(max_init)
-    wp.capture_launch(capture.graph)
-    second_min = out_min.numpy().copy()
-    second_max = out_max.numpy().copy()
-
-    np.testing.assert_array_equal(first_min, second_min)
-    np.testing.assert_array_equal(first_max, second_max)
-
-
-def test_graph_capture_consumed_return_counter(test, device):
-    """Verify consumed-return atomic counters can be captured and replayed."""
-    if device.is_cpu:
-        test.skipTest("Graph capture requires CUDA")
-
-    n = 64
-    rng = np.random.default_rng(202)
-    data_np = rng.random(n, dtype=np.float32)
-
-    data = wp.array(data_np, dtype=wp.float32, device=device)
-    counter = wp.zeros(1, dtype=wp.int32, device=device)
-    output = wp.zeros(n, dtype=wp.float32, device=device)
-
-    wp.launch(counter_kernel, dim=n, inputs=[data, counter], outputs=[output], device=device)
-    counter.zero_()
-    output.zero_()
-
-    with wp.ScopedCapture(device, force_module_load=False) as capture:
-        wp.launch(counter_kernel, dim=n, inputs=[data, counter], outputs=[output], device=device)
-
-    wp.capture_launch(capture.graph)
-    first_count = counter.numpy().copy()
-    first = output.numpy().copy()
-
-    counter.zero_()
-    output.zero_()
-    wp.capture_launch(capture.graph)
-    second_count = counter.numpy().copy()
-    second = output.numpy().copy()
-
-    np.testing.assert_array_equal(first_count, second_count)
-    np.testing.assert_array_equal(first, second)
-
-
-def test_graph_capture_indexed_counter(test, device):
-    """Verify data-dependent consumed-return counters can be captured and replayed."""
-    if device.is_cpu:
-        test.skipTest("Graph capture requires CUDA")
-
-    n = 64
-    bin_count = 4
-    bins_np = (np.arange(n, dtype=np.int32) * 7) % bin_count
-    values_np = np.arange(n, dtype=np.int32)
-
-    values = wp.array(values_np, dtype=wp.int32, device=device)
-    bins = wp.array(bins_np, dtype=wp.int32, device=device)
-    counter = wp.zeros(bin_count, dtype=wp.int32, device=device)
-    output = wp.full((bin_count, n), value=-1, dtype=wp.int32, device=device)
-
-    wp.launch(indexed_counter_kernel, dim=n, inputs=[values, bins, counter], outputs=[output], device=device)
-    counter.zero_()
-    output.fill_(-1)
-
-    with wp.ScopedCapture(device, force_module_load=False) as capture:
-        wp.launch(indexed_counter_kernel, dim=n, inputs=[values, bins, counter], outputs=[output], device=device)
-
-    wp.capture_launch(capture.graph)
-    first_count = counter.numpy().copy()
-    first = output.numpy().copy()
-
-    counter.zero_()
-    output.fill_(-1)
-    wp.capture_launch(capture.graph)
-    second_count = counter.numpy().copy()
-    second = output.numpy().copy()
-
-    np.testing.assert_array_equal(first_count, second_count)
-    np.testing.assert_array_equal(first, second)
-
-
-def test_counter_large_launch_rejected(test, device):
-    """Verify counter prefix buffers fail clearly before oversized launches."""
-    if device.is_cpu:
-        test.skipTest("CUDA deterministic counter path required")
-
-    data = wp.ones(1, dtype=wp.float32, device=device)
-    counter = wp.zeros(1, dtype=wp.int32, device=device)
-    output = wp.zeros(1, dtype=wp.float32, device=device)
-
-    with test.assertRaisesRegex(RuntimeError, "up to 2\\^31 - 1 threads"):
-        wp.launch(counter_kernel, dim=(46341, 46341), inputs=[data, counter], outputs=[output], device=device)
-
-
-def test_apic_capture_rejects_deterministic_cuda_kernel(test, device):
-    """Verify APIC serialization fails explicitly for deterministic CUDA kernels."""
-    if device.is_cpu:
-        test.skipTest("CUDA APIC capture path required")
-
-    n = 16
-    data_np = np.ones(n, dtype=np.float32)
-    indices_np = np.arange(n, dtype=np.int32) % 4
-
-    data = wp.array(data_np, dtype=wp.float32, device=device)
-    indices = wp.array(indices_np, dtype=wp.int32, device=device)
-    output = wp.zeros(4, dtype=wp.float32, device=device)
-
-    wp.launch(scatter_add_kernel, dim=n, inputs=[data, indices], outputs=[output], device=device)
-    output.zero_()
-
-    with test.assertRaisesRegex(RuntimeError, "APIC serialization"):
-        with wp.ScopedCapture(device=device, apic=True, force_module_load=False):
-            wp.launch(scatter_add_kernel, dim=n, inputs=[data, indices], outputs=[output], device=device)
-
-
-def test_deterministic_backward_scatter_add(test, device):
-    """Verify deterministic scatter-add kernels launch backward and propagate value gradients."""
-    if device.is_cpu:
-        test.skipTest("CUDA backward launch coverage required")
-
-    n = 512
-    out_size = 16
-    rng = np.random.default_rng(300)
-    data_np = rng.random(n, dtype=np.float32)
-    indices_np = rng.integers(0, out_size, size=n, dtype=np.int32)
-
-    data = wp.array(data_np, dtype=wp.float32, device=device, requires_grad=True)
-    indices = wp.array(indices_np, dtype=wp.int32, device=device)
-    output = wp.zeros(out_size, dtype=wp.float32, device=device, requires_grad=True)
-
-    tape = wp.Tape()
-    with tape:
-        wp.launch(scatter_add_kernel, dim=n, inputs=[data, indices], outputs=[output], device=device)
-
-    tape.backward(grads={output: wp.ones_like(output)})
-
-    np.testing.assert_allclose(tape.gradients[data].numpy(), np.ones(n, dtype=np.float32), rtol=0, atol=0)
-
-
-def test_deterministic_backward_address_scatter(test, device):
-    """Verify generated array-read adjoints reduce contended gradients deterministically."""
-    if device.is_cpu:
-        test.skipTest("CUDA backward launch coverage required")
-
-    n = 4096
-    value_count = 37
-    rng = np.random.default_rng(303)
-    values_np = rng.random(value_count, dtype=np.float32)
-    indices_np = rng.integers(0, value_count, size=n, dtype=np.int32)
-    grad_np = rng.random(n, dtype=np.float32)
-    expected = _reference_scatter_add_float32(grad_np, indices_np, value_count)
-
-    values = wp.array(values_np, dtype=wp.float32, device=device, requires_grad=True)
-    indices = wp.array(indices_np, dtype=wp.int32, device=device)
-    output_grad = wp.array(grad_np, dtype=wp.float32, device=device)
-
-    old_det = _get_test_module_options()["deterministic"]
-    try:
-        _set_test_module_options({"deterministic": "gpu_to_gpu"})
-        results = []
-        for _ in range(3):
-            values.grad.zero_()
-            output = wp.zeros(n, dtype=wp.float32, device=device, requires_grad=True)
-            tape = wp.Tape()
-            with tape:
-                wp.launch(gather_address_kernel, dim=n, inputs=[values, indices], outputs=[output], device=device)
-            tape.backward(grads={output: output_grad})
-            results.append(tape.gradients[values].numpy().copy())
-    finally:
-        _set_test_module_options({"deterministic": old_det})
-
-    for result in results:
-        np.testing.assert_array_equal(result, expected)
-
-
-def test_deterministic_backward_vec3_address_scatter(test, device):
-    """Verify vector-valued array-read adjoints are also reduced deterministically."""
-    if device.is_cpu:
-        test.skipTest("CUDA backward launch coverage required")
-
-    n = 2048
-    value_count = 29
-    rng = np.random.default_rng(304)
-    values_np = rng.random((value_count, 3), dtype=np.float32)
-    indices_np = rng.integers(0, value_count, size=n, dtype=np.int32)
-    grad_np = rng.random((n, 3), dtype=np.float32)
-    expected = np.zeros((value_count, 3), dtype=np.float32)
-    for value, index in zip(grad_np, indices_np, strict=True):
-        expected[index] = np.float32(expected[index] + value)
-
-    values = wp.array(values_np, dtype=wp.vec3, device=device, requires_grad=True)
-    indices = wp.array(indices_np, dtype=wp.int32, device=device)
-    output_grad = wp.array(grad_np, dtype=wp.vec3, device=device)
-
-    old_det = _get_test_module_options()["deterministic"]
-    try:
-        _set_test_module_options({"deterministic": "gpu_to_gpu"})
-        results = []
-        for _ in range(3):
-            values.grad.zero_()
-            output = wp.zeros(n, dtype=wp.vec3, device=device, requires_grad=True)
-            tape = wp.Tape()
-            with tape:
-                wp.launch(vec3_gather_address_kernel, dim=n, inputs=[values, indices], outputs=[output], device=device)
-            tape.backward(grads={output: output_grad})
-            results.append(tape.gradients[values].numpy().copy())
-    finally:
-        _set_test_module_options({"deterministic": old_det})
-
-    for result in results:
-        np.testing.assert_array_equal(result, expected)
-
-
-def test_deterministic_backward_missing_adjoint_target(test, device):
-    """Verify backward deterministic reductions ignore array reads with no grad buffer."""
-    if device.is_cpu:
-        test.skipTest("CUDA backward launch coverage required")
-
-    n = 2048
-    value_count = 23
-    rng = np.random.default_rng(305)
-    values_np = rng.random(value_count, dtype=np.float32)
-    indices_np = rng.integers(0, value_count, size=n, dtype=np.int32)
-    grad_np = rng.random(n, dtype=np.float32)
-    expected = np.zeros(1, dtype=np.float32)
-    for grad, index in zip(grad_np, indices_np, strict=True):
-        expected[0] = np.float32(expected[0] + np.float32(grad * values_np[index]))
-
-    values = wp.array(values_np, dtype=wp.float32, device=device, requires_grad=False)
-    indices = wp.array(indices_np, dtype=wp.int32, device=device)
-    scale = wp.ones(1, dtype=wp.float32, device=device, requires_grad=True)
-    output_grad = wp.array(grad_np, dtype=wp.float32, device=device)
-
-    old_det = _get_test_module_options()["deterministic"]
-    try:
-        _set_test_module_options({"deterministic": "gpu_to_gpu"})
-        results = []
-        for _ in range(3):
-            scale.grad.zero_()
-            output = wp.zeros(n, dtype=wp.float32, device=device, requires_grad=True)
-            tape = wp.Tape()
-            with tape:
-                wp.launch(
-                    gather_with_nongrad_input_kernel,
-                    dim=n,
-                    inputs=[values, indices, scale],
-                    outputs=[output],
-                    device=device,
-                )
-            tape.backward(grads={output: output_grad})
-            results.append(tape.gradients[scale].numpy().copy())
-    finally:
-        _set_test_module_options({"deterministic": old_det})
-
-    for result in results:
-        np.testing.assert_array_equal(result, expected)
-
-
-def test_deterministic_backward_counter_store(test, device):
-    """Verify counter-mode array stores preserve ``array_store`` adjoints."""
-    if device.is_cpu:
-        test.skipTest("CUDA backward launch coverage required")
-
-    n = 64
-    data_np = np.arange(n, dtype=np.float32)
-
-    data = wp.array(data_np, dtype=wp.float32, device=device, requires_grad=True)
-    counter = wp.zeros(1, dtype=wp.int32, device=device)
-    output = wp.zeros(n, dtype=wp.float32, device=device, requires_grad=True)
-
-    tape = wp.Tape()
-    with tape:
-        wp.launch(counter_kernel, dim=n, inputs=[data, counter], outputs=[output], device=device)
-
-    tape.backward(grads={output: wp.ones_like(output)})
-
-    np.testing.assert_allclose(tape.gradients[data].numpy(), np.ones(n, dtype=np.float32), rtol=0, atol=0)
-
-
-def test_deterministic_custom_replay_counter(test, device):
-    """Verify deterministic helper args do not leak into custom replay calls."""
-    if device.is_cpu:
-        test.skipTest("CUDA custom replay coverage required")
-
-    n = 64
-    data_np = np.arange(n, dtype=np.float32)
-
-    data = wp.array(data_np, dtype=wp.float32, device=device, requires_grad=True)
-    counter = wp.zeros(1, dtype=wp.int32, device=device)
-    tids = wp.zeros(n, dtype=wp.int32, device=device)
-    output = wp.zeros(n, dtype=wp.float32, device=device, requires_grad=True)
-
-    tape = wp.Tape()
-    with tape:
-        wp.launch(custom_replay_counter_kernel, dim=n, inputs=[data, counter, tids], outputs=[output], device=device)
-
-    tape.backward(grads={output: wp.ones_like(output)})
-
-    np.testing.assert_allclose(tape.gradients[data].numpy(), 2.0 * data_np, rtol=0, atol=0)
-
-
-def test_deterministic_custom_adjoint_array_atomic(test, device):
-    """Verify custom adjoints that atomically update ``wp.adjoint[array]`` compile."""
-    if device.is_cpu:
-        test.skipTest("CUDA backward launch coverage required")
-
-    n = 48
-    value_count = 7
-    indices_np = (np.arange(n, dtype=np.int32) * 3) % value_count
-    values_np = np.linspace(0.25, 1.75, value_count, dtype=np.float32)
-
-    values = wp.array(values_np, dtype=wp.float32, device=device, requires_grad=True)
-    indices = wp.array(indices_np, dtype=wp.int32, device=device)
-    output = wp.zeros(1, dtype=wp.float32, device=device, requires_grad=True)
-
-    tape = wp.Tape()
-    with tape:
-        wp.launch(custom_adjoint_lookup_kernel, dim=n, inputs=[values, indices], outputs=[output], device=device)
-
-    tape.backward(grads={output: wp.ones_like(output)})
-
-    expected = np.bincount(indices_np, minlength=value_count).astype(np.float32)
-    np.testing.assert_allclose(tape.gradients[values].numpy(), expected, rtol=0, atol=0)
-
-
-def test_deterministic_custom_adjoint_gather_atomic(test, device):
-    """Verify custom ``wp.adjoint[array]`` atomics use deterministic backward reduction."""
-    if device.is_cpu:
-        test.skipTest("CUDA backward launch coverage required")
-
-    n = 2048
-    value_count = 31
-    rng = np.random.default_rng(305)
-    values_np = rng.random(value_count, dtype=np.float32)
-    indices_np = rng.integers(0, value_count, size=n, dtype=np.int32)
-    grad_np = rng.random(n, dtype=np.float32)
-    expected = _reference_scatter_add_float32(grad_np, indices_np, value_count)
-
-    values = wp.array(values_np, dtype=wp.float32, device=device, requires_grad=True)
-    indices = wp.array(indices_np, dtype=wp.int32, device=device)
-    output_grad = wp.array(grad_np, dtype=wp.float32, device=device)
-
-    old_det = _get_test_module_options()["deterministic"]
-    try:
-        _set_test_module_options({"deterministic": "gpu_to_gpu"})
-        results = []
-        for _ in range(3):
-            values.grad.zero_()
-            output = wp.zeros(n, dtype=wp.float32, device=device, requires_grad=True)
-            tape = wp.Tape()
-            with tape:
-                wp.launch(
-                    custom_adjoint_gather_kernel, dim=n, inputs=[values, indices], outputs=[output], device=device
-                )
-            tape.backward(grads={output: output_grad})
-            results.append(tape.gradients[values].numpy().copy())
-    finally:
-        _set_test_module_options({"deterministic": old_det})
-
-    for result in results:
-        np.testing.assert_array_equal(result, expected)
-
-
-def test_deterministic_custom_adjoint_store_function(test, device):
-    """Verify custom adjoints on functions with deterministic store context compile."""
-    if device.is_cpu:
-        test.skipTest("CUDA backward launch coverage required")
-
-    n = 32
-    values_np = np.linspace(0.5, 2.0, n, dtype=np.float32)
-
-    values = wp.array(values_np, dtype=wp.float32, device=device, requires_grad=True)
-    output = wp.zeros(n, dtype=wp.float32, device=device, requires_grad=True)
-    scratch = wp.ones(n, dtype=wp.float32, device=device)
-
-    tape = wp.Tape()
-    with tape:
-        wp.launch(custom_adjoint_store_kernel, dim=n, inputs=[values], outputs=[output, scratch], device=device)
-
-    tape.backward(grads={output: wp.ones_like(output)})
-
-    np.testing.assert_allclose(output.numpy(), values_np * values_np, rtol=0, atol=0)
-    np.testing.assert_allclose(scratch.numpy(), np.zeros(n, dtype=np.float32), rtol=0, atol=0)
-    np.testing.assert_allclose(tape.gradients[values].numpy(), 2.0 * values_np, rtol=0, atol=0)
-
-
-def test_custom_adjoint_not_guaranteed_mode(test, device):
-    """Verify custom adjoints compile after disabling deterministic mode."""
-    n = 48
-    value_count = 7
-    indices_np = (np.arange(n, dtype=np.int32) * 3) % value_count
-    values_np = np.linspace(0.25, 1.75, value_count, dtype=np.float32)
-
-    old_det = _get_test_module_options()["deterministic"]
-    try:
-        _set_test_module_options({"deterministic": "not_guaranteed"})
-
-        values = wp.array(values_np, dtype=wp.float32, device=device, requires_grad=True)
-        indices = wp.array(indices_np, dtype=wp.int32, device=device)
-        output = wp.zeros(n, dtype=wp.float32, device=device, requires_grad=True)
-
-        tape = wp.Tape()
-        with tape:
-            wp.launch(
-                custom_adjoint_not_guaranteed_kernel, dim=n, inputs=[values, indices], outputs=[output], device=device
-            )
-
-        tape.backward(grads={output: wp.ones_like(output)})
-        result = tape.gradients[values].numpy()
-
-        if not device.is_cpu:
-            _set_test_module_options({"deterministic": "run_to_run"})
-
-            store_n = 32
-            store_values_np = np.linspace(0.5, 2.0, store_n, dtype=np.float32)
-            store_values = wp.array(store_values_np, dtype=wp.float32, device=device, requires_grad=True)
-            store_output = wp.zeros(store_n, dtype=wp.float32, device=device, requires_grad=True)
-            store_scratch = wp.ones(store_n, dtype=wp.float32, device=device)
-
-            store_tape = wp.Tape()
-            with store_tape:
-                wp.launch(
-                    custom_adjoint_store_kernel,
-                    dim=store_n,
-                    inputs=[store_values],
-                    outputs=[store_output, store_scratch],
-                    device=device,
-                )
-
-            store_tape.backward(grads={store_output: wp.ones_like(store_output)})
-            store_output_result = store_output.numpy()
-            store_scratch_result = store_scratch.numpy()
-            store_gradient_result = store_tape.gradients[store_values].numpy()
-    finally:
-        _set_test_module_options({"deterministic": old_det})
-
-    expected = np.bincount(indices_np, minlength=value_count).astype(np.float32)
-    np.testing.assert_allclose(result, expected, rtol=0, atol=0)
-    if not device.is_cpu:
-        np.testing.assert_allclose(store_output_result, store_values_np * store_values_np, rtol=0, atol=0)
-        np.testing.assert_allclose(store_scratch_result, np.zeros(store_n, dtype=np.float32), rtol=0, atol=0)
-        np.testing.assert_allclose(store_gradient_result, 2.0 * store_values_np, rtol=0, atol=0)
-
-
-def test_deterministic_enum_parity(test, device):
-    """Keep Python deterministic constants aligned with the native enums."""
-    del device
-
-    native_source = (Path(wp.__file__).resolve().parent / "native" / "deterministic.cu").read_text()
-
-    def parse_enum(enum_name):
-        match = re.search(rf"enum {enum_name} \{{(.*?)\n\}};", native_source, re.DOTALL)
-        if match is None:
-            raise AssertionError(f"Failed to find enum {enum_name} in deterministic.cu")
-
-        entries = {}
-        for enum_line in match.group(1).splitlines():
-            entry = enum_line.strip().rstrip(",")
-            if not entry or "=" not in entry:
-                continue
-            name, value = (part.strip() for part in entry.split("=", 1))
-            if name.isidentifier() and value.isdecimal():
-                entries[name] = int(value)
-        return entries
-
-    native_reduce_ops = parse_enum("ReduceOp")
-    native_deterministic_levels = parse_enum("DeterminismLevel")
-    native_scalar_types = parse_enum("ScalarType")
-
-    test.assertEqual(
-        native_reduce_ops,
-        {
-            "REDUCE_OP_ADD": wp_deterministic.REDUCE_OP_ADD,
-            "REDUCE_OP_MIN": wp_deterministic.REDUCE_OP_MIN,
-            "REDUCE_OP_MAX": wp_deterministic.REDUCE_OP_MAX,
-        },
-    )
-    test.assertEqual(
-        native_deterministic_levels,
-        {
-            "DETERMINISTIC_NOT_GUARANTEED": wp_deterministic._DETERMINISTIC_MODE_IDS[
-                wp_deterministic.DETERMINISTIC_NOT_GUARANTEED
-            ],
-            "DETERMINISTIC_RUN_TO_RUN": wp_deterministic._DETERMINISTIC_MODE_IDS[
-                wp_deterministic.DETERMINISTIC_RUN_TO_RUN
-            ],
-            "DETERMINISTIC_GPU_TO_GPU": wp_deterministic._DETERMINISTIC_MODE_IDS[
-                wp_deterministic.DETERMINISTIC_GPU_TO_GPU
-            ],
-        },
-    )
-    test.assertEqual(
-        native_scalar_types,
-        {
-            "SCALAR_HALF": wp_deterministic._SCALAR_TYPE_IDS[wp.float16],
-            "SCALAR_FLOAT": wp_deterministic._SCALAR_TYPE_IDS[wp.float32],
-            "SCALAR_DOUBLE": wp_deterministic._SCALAR_TYPE_IDS[wp.float64],
-            "SCALAR_INT": wp_deterministic._SCALAR_TYPE_IDS[wp.int32],
-            "SCALAR_UINT": wp_deterministic._SCALAR_TYPE_IDS[wp.uint32],
-            "SCALAR_INT64": wp_deterministic._SCALAR_TYPE_IDS[wp.int64],
-            "SCALAR_UINT64": wp_deterministic._SCALAR_TYPE_IDS[wp.uint64],
-            "SCALAR_BFLOAT16": wp_deterministic._SCALAR_TYPE_IDS[wp.bfloat16],
-        },
-    )
-
-
-def test_run_sort_reduce_uses_emitted_record_count(test, device):
-    """Verify the postpass uses emitted records instead of full scatter capacity."""
-    del device
-
-    class FakeCore:
-        def __init__(self):
-            self.workspace_counts = []
-            self.device_counts = []
-
-        def wp_deterministic_sort_reduce_workspace_size(
-            self, count, _op, _scalar_type, _components, _determinism_level
-        ):
-            self.workspace_counts.append(count)
-            return 1
-
-        def wp_deterministic_sort_reduce_device(
-            self,
-            _keys,
-            _values,
-            count,
-            _dest_array,
-            _dest_size,
-            _op,
-            _scalar_type,
-            _components,
-            _determinism_level,
-            _workspace,
-            _workspace_size,
-        ):
-            self.device_counts.append(count)
-
-    class FakeRuntime:
-        def __init__(self):
-            self.core = FakeCore()
-
-    runtime = FakeRuntime()
-    device = wp.get_device("cpu")
-    keys = wp.empty(1024, dtype=wp.int64, device=device)
-    values = wp.empty(1024, dtype=wp.float32, device=device)
-    counter = wp.array(np.array([3], dtype=np.int32), dtype=wp.int32, device=device)
-    dest = wp.zeros(1, dtype=wp.float32, device=device)
-    target = wp_deterministic.ScatterTarget(
-        array_var_label="dest",
-        helper_name="det_dest",
-        family=wp_deterministic.DETERMINISTIC_FAMILY_ADD,
-        value_dtype=wp.float32,
-        value_ctype="float",
-        scalar_dtype=wp.float32,
-        reduce_op=wp_deterministic.REDUCE_OP_ADD,
-    )
-
-    record_count = wp_deterministic.get_scatter_record_count((keys, values, counter, 1024))
-    workspaces = wp_deterministic.run_sort_reduce(
-        runtime,
-        [target],
-        [(keys, values, counter, 1024)],
-        [dest],
-        device,
-        wp_deterministic.DETERMINISTIC_RUN_TO_RUN,
-        record_counts=[record_count],
-    )
-
-    test.assertEqual(record_count, 3)
-    test.assertEqual(runtime.core.workspace_counts, [3])
-    test.assertEqual(runtime.core.device_counts, [3])
-    test.assertEqual(len(workspaces), 1)
-
-
-def test_scatter_record_count_capture_uses_capacity(test, device):
-    """Verify graph-capture launches avoid host-side scatter count readbacks."""
-    del device
-    record_count = wp_deterministic.get_scatter_record_count((None, None, None, 1024), stream_is_capturing=True)
-    test.assertEqual(record_count, 1024)
-
-
-# ---------------------------------------------------------------------------
-# Test class registration
-# ---------------------------------------------------------------------------
-
-cuda_devices = get_selected_cuda_test_devices()
-bfloat16_cuda_devices = [device for device in cuda_devices if device.arch >= 80]
-all_devices = get_test_devices()
-
-
-class TestDeterministic(unittest.TestCase):
-    """Test suite for deterministic execution mode."""
-
-    @classmethod
-    def setUpClass(cls):
-        cls._old_deterministic = wp.config.deterministic
-        cls._old_module_deterministic = _get_test_module_options()["deterministic"]
-        wp.config.deterministic = "run_to_run"
-        _set_test_module_options({"deterministic": "run_to_run"})
-
-    @classmethod
-    def tearDownClass(cls):
-        _set_test_module_options({"deterministic": cls._old_module_deterministic})
-        wp.config.deterministic = cls._old_deterministic
-
-
-# Scatter/reduce atomic tests.
-add_function_test(
-    TestDeterministic, "test_scatter_add_reproducibility", test_scatter_add_reproducibility, devices=cuda_devices
-)
-add_function_test(
-    TestDeterministic,
-    "test_gpu_to_gpu_mode_reproducibility",
-    test_gpu_to_gpu_mode_reproducibility,
-    devices=cuda_devices,
-)
-add_function_test(
-    TestDeterministic,
-    "test_gpu_to_gpu_matches_canonical_float32_reference",
-    test_gpu_to_gpu_matches_canonical_float32_reference,
-    devices=cuda_devices,
-)
-add_function_test(
-    TestDeterministic, "test_augassign_add_reproducibility", test_augassign_add_reproducibility, devices=cuda_devices
-)
-add_function_test(TestDeterministic, "test_scatter_add_correctness", test_scatter_add_correctness, devices=all_devices)
-add_function_test(TestDeterministic, "test_multi_array_atomic", test_multi_array_atomic, devices=cuda_devices)
-add_function_test(
-    TestDeterministic, "test_atomic_sub_deterministic", test_atomic_sub_deterministic, devices=cuda_devices
-)
-add_function_test(TestDeterministic, "test_atomic_add_2d", test_atomic_add_2d, devices=cuda_devices)
-add_function_test(
-    TestDeterministic, "test_sliced_2d_array_atomic_add", test_sliced_2d_array_atomic_add, devices=cuda_devices
-)
-add_function_test(
-    TestDeterministic, "test_sliced_3d_array_atomic_add", test_sliced_3d_array_atomic_add, devices=cuda_devices
-)
-add_function_test(
-    TestDeterministic, "test_strided_1d_view_atomic_add", test_strided_1d_view_atomic_add, devices=cuda_devices
-)
-add_function_test(
-    TestDeterministic, "test_zero_stride_view_atomic_add", test_zero_stride_view_atomic_add, devices=cuda_devices
-)
-add_function_test(TestDeterministic, "test_column_slice_atomic_add", test_column_slice_atomic_add, devices=cuda_devices)
-add_function_test(
-    TestDeterministic, "test_transposed_2d_atomic_add", test_transposed_2d_atomic_add, devices=cuda_devices
-)
-add_function_test(
-    TestDeterministic, "test_atomic_half_deterministic", test_atomic_half_deterministic, devices=cuda_devices
-)
-add_function_test(
-    TestDeterministic,
-    "test_atomic_bfloat16_deterministic",
-    test_atomic_bfloat16_deterministic,
-    devices=bfloat16_cuda_devices,
-    check_output=False,
-)
-add_function_test(
-    TestDeterministic, "test_atomic_double_deterministic", test_atomic_double_deterministic, devices=cuda_devices
-)
-add_function_test(
-    TestDeterministic, "test_vec3_atomic_add_deterministic", test_vec3_atomic_add_deterministic, devices=cuda_devices
-)
-add_function_test(
-    TestDeterministic,
-    "test_vec3_atomic_minmax_deterministic",
-    test_vec3_atomic_minmax_deterministic,
-    devices=cuda_devices,
-)
-add_function_test(
-    TestDeterministic, "test_mat33_atomic_add_deterministic", test_mat33_atomic_add_deterministic, devices=cuda_devices
-)
-add_function_test(
-    TestDeterministic,
-    "test_triple_scatter_capacity_estimate",
-    test_triple_scatter_capacity_estimate,
-    devices=cuda_devices,
-)
-add_function_test(
-    TestDeterministic,
-    "test_loop_scatter_max_records_override",
-    test_loop_scatter_max_records_override,
-    devices=cuda_devices,
-)
-add_function_test(
-    TestDeterministic,
-    "test_scatter_overflow_reports_error",
-    test_scatter_overflow_reports_error,
-    devices=cuda_devices,
-)
-add_function_test(
-    TestDeterministic,
-    "test_scatter_large_launch_rejected",
-    test_scatter_large_launch_rejected,
-    devices=cuda_devices,
-)
-add_function_test(
-    TestDeterministic, "test_mixed_reduce_ops_same_array", test_mixed_reduce_ops_same_array, devices=cuda_devices
-)
-add_function_test(
-    TestDeterministic, "test_deterministic_func_kernel", test_deterministic_func_kernel, devices=cuda_devices
-)
-add_function_test(
-    TestDeterministic,
-    "test_nested_deterministic_func_kernel",
-    test_nested_deterministic_func_kernel,
-    devices=cuda_devices,
-)
-add_function_test(
-    TestDeterministic, "test_struct_field_counter_atomic", test_struct_field_counter_atomic, devices=cuda_devices
-)
-add_function_test(
-    TestDeterministic,
-    "test_struct_field_helper_counter_atomic",
-    test_struct_field_helper_counter_atomic,
-    devices=cuda_devices,
-)
-add_function_test(
-    TestDeterministic,
-    "test_metadata_refresh_struct_tile_stack",
-    test_metadata_refresh_struct_tile_stack,
-    devices=[wp.get_device("cpu")],
-)
-add_function_test(
-    TestDeterministic,
-    "test_config_deterministic_max_records_default",
-    test_config_deterministic_max_records_default,
-    devices=[wp.get_device("cpu")],
-)
-add_function_test(
-    TestDeterministic,
-    "test_helper_name_collision",
-    test_helper_name_collision,
-    devices=cuda_devices,
-)
-
-# Consumed-return counter tests.
-add_function_test(TestDeterministic, "test_counter_reproducibility", test_counter_reproducibility, devices=cuda_devices)
-add_function_test(
-    TestDeterministic,
-    "test_counter_phase0_suppresses_array_writes",
-    test_counter_phase0_suppresses_array_writes,
-    devices=cuda_devices,
-)
-add_function_test(
-    TestDeterministic,
-    "test_counter_phase0_suppresses_array_writes_before_helper",
-    test_counter_phase0_suppresses_array_writes_before_helper,
-    devices=cuda_devices,
-)
-add_function_test(
-    TestDeterministic,
-    "test_counter_phase0_suppresses_helper_array_writes",
-    test_counter_phase0_suppresses_helper_array_writes,
-    devices=cuda_devices,
-)
-add_function_test(
-    TestDeterministic,
-    "test_counter_phase0_preserves_local_scratch_writes",
-    test_counter_phase0_preserves_local_scratch_writes,
-    devices=cuda_devices,
-)
-add_function_test(
-    TestDeterministic,
-    "test_counter_phase0_skips_unintercepted_atomic_xor",
-    test_counter_phase0_skips_unintercepted_atomic_xor,
-    devices=cuda_devices,
-)
-add_function_test(
-    TestDeterministic,
-    "test_counter_phase0_skips_atomic_xor_in_helper_func",
-    test_counter_phase0_skips_atomic_xor_in_helper_func,
-    devices=cuda_devices,
-)
-add_function_test(
-    TestDeterministic,
-    "test_counter_consumed_bitwise_atomic_rejected",
-    test_counter_consumed_bitwise_atomic_rejected,
-    devices=cuda_devices,
-)
-add_function_test(
-    TestDeterministic,
-    "test_counter_phase0_suppresses_component_stores",
-    test_counter_phase0_suppresses_component_stores,
-    devices=cuda_devices,
-)
-add_function_test(TestDeterministic, "test_counter_correctness", test_counter_correctness, devices=all_devices)
-add_function_test(
-    TestDeterministic,
-    "test_counter_nonzero_initial_value",
-    test_counter_nonzero_initial_value,
-    devices=cuda_devices,
-)
-add_function_test(
-    TestDeterministic,
-    "test_counter_multi_launch_accumulates",
-    test_counter_multi_launch_accumulates,
-    devices=cuda_devices,
-)
-add_function_test(
-    TestDeterministic,
-    "test_counter_variable_total_writeback",
-    test_counter_variable_total_writeback,
-    devices=cuda_devices,
-)
-add_function_test(
-    TestDeterministic,
-    "test_counter_nonzero_index",
-    test_counter_nonzero_index,
-    devices=cuda_devices,
-)
-add_function_test(
-    TestDeterministic,
-    "test_counter_indexed_destinations",
-    test_counter_indexed_destinations,
-    devices=cuda_devices,
-)
-add_function_test(
-    TestDeterministic,
-    "test_counter_sliced_destinations",
-    test_counter_sliced_destinations,
-    devices=cuda_devices,
-)
-add_function_test(
-    TestDeterministic,
-    "test_counter_indexed_dynamic_loop",
-    test_counter_indexed_dynamic_loop,
-    devices=cuda_devices,
-)
-add_function_test(
-    TestDeterministic,
-    "test_counter_int64_rejected",
-    test_counter_int64_rejected,
-    devices=cuda_devices,
-)
-add_function_test(
-    TestDeterministic,
-    "test_counter_non_add_atomic_rejected",
-    test_counter_non_add_atomic_rejected,
-    devices=cuda_devices,
-)
-add_function_test(
-    TestDeterministic,
-    "test_float_atomic_consumed_return_rejected",
-    test_float_atomic_consumed_return_rejected,
-    devices=cuda_devices,
-)
-add_function_test(
-    TestDeterministic,
-    "test_float_atomic_bare_statement_allowed",
-    test_float_atomic_bare_statement_allowed,
-    devices=cuda_devices,
-)
-add_function_test(
-    TestDeterministic,
-    "test_scatter_capacity_overflow_rejected",
-    test_scatter_capacity_overflow_rejected,
-    devices=[wp.get_device("cpu")],
-)
-add_function_test(TestDeterministic, "test_conditional_counter", test_conditional_counter, devices=cuda_devices)
-
-# Mixed counter and scatter/reduce tests.
-add_function_test(TestDeterministic, "test_mixed_pattern", test_mixed_pattern, devices=cuda_devices)
-add_function_test(
-    TestDeterministic,
-    "test_counter_with_integer_accumulation",
-    test_counter_with_integer_accumulation,
-    devices=cuda_devices,
-)
-
-# Passthrough / override tests.
-add_function_test(TestDeterministic, "test_int_atomic_passthrough", test_int_atomic_passthrough, devices=all_devices)
-add_function_test(TestDeterministic, "test_module_option_override", test_module_option_override, devices=all_devices)
-add_function_test(
-    TestDeterministic, "test_deterministic_closure_kernel", test_deterministic_closure_kernel, devices=cuda_devices
-)
-add_function_test(
-    TestDeterministic,
-    "test_record_cmd_deterministic_launch",
-    test_record_cmd_deterministic_launch,
-    devices=cuda_devices,
-)
-add_function_test(
-    TestDeterministic,
-    "test_graph_capture_deterministic_launch",
-    test_graph_capture_deterministic_launch,
-    devices=cuda_devices,
-)
-add_function_test(
-    TestDeterministic,
-    "test_graph_capture_sliced_array",
-    test_graph_capture_sliced_array,
-    devices=cuda_devices,
-)
-add_function_test(
-    TestDeterministic,
-    "test_graph_capture_deterministic_closure_kernel",
-    test_graph_capture_deterministic_closure_kernel,
-    devices=cuda_devices,
-)
-add_function_test(
-    TestDeterministic,
-    "test_graph_capture_deterministic_func_kernel",
-    test_graph_capture_deterministic_func_kernel,
-    devices=cuda_devices,
-)
-add_function_test(
-    TestDeterministic,
-    "test_graph_capture_vec3_atomic_minmax",
-    test_graph_capture_vec3_atomic_minmax,
-    devices=cuda_devices,
-)
-add_function_test(
-    TestDeterministic,
-    "test_graph_capture_consumed_return_counter",
-    test_graph_capture_consumed_return_counter,
-    devices=cuda_devices,
-)
-add_function_test(
-    TestDeterministic,
-    "test_graph_capture_indexed_counter",
-    test_graph_capture_indexed_counter,
-    devices=cuda_devices,
-)
-add_function_test(
-    TestDeterministic,
-    "test_counter_large_launch_rejected",
-    test_counter_large_launch_rejected,
-    devices=cuda_devices,
-)
-add_function_test(
-    TestDeterministic,
-    "test_apic_capture_rejects_deterministic_cuda_kernel",
-    test_apic_capture_rejects_deterministic_cuda_kernel,
-    devices=cuda_devices,
-)
-add_function_test(
-    TestDeterministic,
-    "test_deterministic_backward_scatter_add",
-    test_deterministic_backward_scatter_add,
-    devices=cuda_devices,
-)
-add_function_test(
-    TestDeterministic,
-    "test_deterministic_backward_address_scatter",
-    test_deterministic_backward_address_scatter,
-    devices=cuda_devices,
-)
-add_function_test(
-    TestDeterministic,
-    "test_deterministic_backward_vec3_address_scatter",
-    test_deterministic_backward_vec3_address_scatter,
-    devices=cuda_devices,
-)
-add_function_test(
-    TestDeterministic,
-    "test_deterministic_backward_missing_adjoint_target",
-    test_deterministic_backward_missing_adjoint_target,
-    devices=cuda_devices,
-)
-add_function_test(
-    TestDeterministic,
-    "test_deterministic_backward_counter_store",
-    test_deterministic_backward_counter_store,
-    devices=cuda_devices,
-)
-add_function_test(
-    TestDeterministic,
-    "test_deterministic_custom_replay_counter",
-    test_deterministic_custom_replay_counter,
-    devices=cuda_devices,
-)
-add_function_test(
-    TestDeterministic,
-    "test_deterministic_custom_adjoint_array_atomic",
-    test_deterministic_custom_adjoint_array_atomic,
-    devices=cuda_devices,
-)
-add_function_test(
-    TestDeterministic,
-    "test_deterministic_custom_adjoint_gather_atomic",
-    test_deterministic_custom_adjoint_gather_atomic,
-    devices=cuda_devices,
-)
-add_function_test(
-    TestDeterministic,
-    "test_deterministic_custom_adjoint_store_function",
-    test_deterministic_custom_adjoint_store_function,
-    devices=cuda_devices,
-)
-add_function_test(
-    TestDeterministic,
-    "test_custom_adjoint_not_guaranteed_mode",
-    test_custom_adjoint_not_guaranteed_mode,
-    devices=all_devices,
-)
-add_function_test(
-    TestDeterministic, "test_deterministic_enum_parity", test_deterministic_enum_parity, devices=all_devices
-)
-add_function_test(
-    TestDeterministic,
-    "test_run_sort_reduce_uses_emitted_record_count",
-    test_run_sort_reduce_uses_emitted_record_count,
-    devices=[wp.get_device("cpu")],
-)
-add_function_test(
-    TestDeterministic,
-    "test_scatter_record_count_capture_uses_capacity",
-    test_scatter_record_count_capture_uses_capacity,
-    devices=[wp.get_device("cpu")],
-)
-
-
-if __name__ == "__main__":
-    unittest.main(verbosity=2)
